% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{assign_cluster_id}
\alias{assign_cluster_id}
\title{Assign Cluster ID to features}
\usage{
assign_cluster_id(clusters, features, name_col)
}
\arguments{
\item{clusters}{a list of clusters as returned by find_clusters}

\item{features}{data frame with feature information, fData(object)}

\item{name_col}{character, name of the column in features that contains feature names}
}
\description{
Assings a cluster ID to all features that are part of a cluster with 2 or more features.
}
