% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{clean_stats_results}
\alias{clean_stats_results}
\title{Statistics cleaning}
\usage{
clean_stats_results(
  df,
  remove = c("Intercept", "CI", "Std_error", "t_value", "z_value", "R2"),
  rename = NULL,
  summary = TRUE,
  plots = TRUE,
  p_limit = 0.05,
  fdr = TRUE
)
}
\arguments{
\item{df}{data frame, statistics results}

\item{remove}{list, should contain strings that are matching to unwanted columns}

\item{rename}{Deprecated}

\item{summary}{logical, should summary columns be added}

\item{plots}{logical, should histograms of p-values be plotted}

\item{p_limit}{numeric, limit for p-values to be counted}

\item{fdr}{logical, should summary be done with fdr-fixed values}
}
\description{
Uses regular expression to remove unnecessary columns from statistics results data frame.
Calculates how many significant p-values are below a certain limit and which is the lowest p-value
and reports the columns. Defaults to use FDR fixed p-values in the summary. Plots histograms of p-values.
}
\examples{
# Simple manipulation to linear model results
lm_results <- perform_lm(drop_qcs(example_set), formula_char = "Feature ~ Group * Time")
lm_results <- clean_stats_results(lm_results)

}
