% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cles}
\alias{cles}
\title{Common Language Effect Size}
\usage{
cles(object, group = group_col(object), id = NULL, separate_by = NULL)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the group column}

\item{id}{character, name of the subject ID column for the paired version.
If \code{NULL}, the test is assumed to be between groups.}

\item{separate_by}{character, name of the column on which you want to
separate the tests. Usually used when dealing with multiple sample types
in the same object.}
}
\description{
Computes Common Language Effect Size (CLES) for each feature.
Calculates CLES using U statistic from Mann-Whitney u-test.
}
\examples{
# Basic example
cles <- cles(drop_qcs(example_set))

# Separately to each time point in each group
cles <- cles(drop_qcs(example_set), group = "Time", separate_by = "Group")
}
