% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{common_language_effect_size}
\alias{common_language_effect_size}
\title{Common Language Effect Size}
\usage{
common_language_effect_size(
  object,
  group = group_col(object),
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the group column}

\item{separate_by}{character, name of the column on which you want to
separate the tests. Usually used when dealing with multiple sample types
in the same object.}
}
\description{
Computes Common Language Effect Size (CLES) for each feature.
Calculates CLES using U statistic from Mann-Whitney u-test. Therefore, the effect size is probability
that a randomly selected observation from the first level of the grouping variable is larger than a randomly
selected observation from the second level of the grouping variable.
}
\examples{
# Basic example
cles <- common_language_effect_size(drop_qcs(example_set))
# Separately to each time point in each group
cles <- common_language_effect_size(drop_qcs(example_set), group = "Time", separate_by = "Group")
}
