% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{construct_MetaboSet}
\alias{construct_MetaboSet}
\title{Construct MetaboSet objects}
\usage{
construct_MetaboSet(exprs, pheno_data, feature_data,
  group_col = NA_character_, time_col = NA_character_,
  subject_col = NA_character_)
}
\arguments{
\item{exprs}{matrix, the feature abundances, size features x samples}

\item{pheno_data}{data frame, sample information, size sample info variables x samples}

\item{feature_data}{data frame, information about the features, size features x feature info columns}

\item{group_col}{character, the name of the column in pheno_data to use as the grouping variable}

\item{time_col}{character, the name of the column in pheno_data to use as the time variable}

\item{subject_col}{character, the name of the column in pheno_data to use as the subject ID variable}
}
\value{
list of MetaboSet objects
}
\description{
Construct MetaboSet objects from input read by read_from_excel.
Returns a list of MetaboSet objects, one per mode. The modes are separated by the "Split" column
in feature_data.
}
\seealso{
\code{\link{read_from_excel}}
}
