% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{correct_drift}
\alias{correct_drift}
\title{Correct drift}
\usage{
correct_drift(object, spar = NULL, spar_lower = 0.5,
  spar_upper = 1.5)
}
\arguments{
\item{object}{a MetaboSet object}

\item{spar}{smoothing parameter}

\item{spar_lower, spar_upper}{lower and upper limits for the smoothing parameter}
}
\value{
MetaboSet object as the one supplied, with drift corrected fetures
}
\description{
Corrects the dirft in the features by applying smoothed cubic spline regression
to each feature separately.
}
\details{
If \code{spar} is set to \code{NULL} (the default), the smoothing parameter will
be separately chosen for each feature from the range [\code{spar_lower, spar_upper}]
using cross validation.
}
\seealso{
\code{\link[stats]{smooth.spline}} for details about the regression,
\code{\link{inspect_dc}} for analysing the drift correction results,
\code{\link{plot_dc}} for plotting the drift correction process for each feature

\code{\link{inspect_dc}}, \code{\link{save_dc_plots}}
}
