% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{dobc}
\alias{dobc}
\title{Batch correction}
\usage{
dobc(object, batch, ref, ref_label, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch}{the column name for batch labels}

\item{ref}{the column name for reference sample labels}

\item{ref_label}{the label for reference samples}

\item{...}{other parameters pased to doBC}
}
\description{
DEPRECATED
}
