% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onLoad}
\alias{.onLoad}
\title{Set default color scales on load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname, pkgname}{default parameters}
}
\description{
Set default color scales on load
}
