% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{drop_flagged}
\alias{drop_flagged}
\title{Drop flagged features}
\usage{
drop_flagged(object, all_features = FALSE)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be retained? Mainy used by internal functions}
}
\description{
Removes all features that have been flagged by quality control functions.
Only features that do not have a flag (Flag == NA) are retained.
}
