% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{drop_qcs}
\alias{drop_qcs}
\title{Drop QC samples}
\usage{
drop_qcs(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
MetaboSet object as the one supplied, withou QC samples
}
\description{
Drop QC samples
}
