% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_set}
\alias{example_set}
\title{Toy data set}
\format{
An object of class \code{MetaboSet} with 20 rows and 30 columns.
}
\usage{
example_set
}
\description{
Contains imaginary data used in testing the package functions.
For more realistic data, see \code{\link{merged_sample}}
This dataset includes multiple observations from same subjects,
sampled at two timepoints and divided to two groups.
The dataset has 30 samples and 20 features.
}
\keyword{datasets}
