% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{finish_log}
\alias{finish_log}
\title{Finish a log}
\usage{
finish_log()
}
\value{
None, the function is invoked for its side effect.
}
\description{
Logs the current date and time and session info, and switches logging off.
}
\examples{
finish_log()

}
\seealso{
\code{\link{init_log}}, \code{\link{log_text}}
}
