% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finite_helpers}
\alias{finite_helpers}
\alias{finite_sd}
\alias{finite_mean}
\alias{finite_median}
\alias{finite_min}
\alias{finite_max}
\alias{finite_mad}
\alias{finite_quantile}
\title{Summary statistics of finite elements}
\usage{
finite_sd(x)

finite_mean(x)

finite_median(x)

finite_min(x)

finite_max(x)

finite_mad(x)

finite_quantile(x, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{other parameters passed to underlying function}
}
\value{
A named, numeric vector with the summary statistic in question.
}
\description{
These functions first remove non-finite and missing values, then compute the 
summary statistic in question.
They are helper functions used for computing quality measurements.
}
\examples{
data <- combined_data(example_set)
features <- fData(example_set)
features$MPA <- sapply(data[, features[, "Feature_ID"]], finite_median)

}
