% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{flag}
\alias{flag}
\alias{flag<-}
\title{These generic functions access and set the values in the flag column}
\usage{
flag(object)

flag(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{character vector, values for flag column}
}
\description{
These generic functions access and set the values in the flag column
}
