% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{flag_report}
\alias{flag_report}
\title{A report of flagged features}
\usage{
flag_report(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
data frame with the number of features at each stage of flagging
}
\description{
Computes the number of features at each stage of flagging for each mode.
}
\examples{
flagged <- merged_sample \%>\%
  mark_nas(0) \%>\%
  flag_detection() \%>\%
  flag_quality()
flag_report(flagged)

}
