% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{impute_simple}
\alias{impute_simple}
\title{Simple imputation}
\usage{
impute_simple(object, value, na_limit = 0)
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{the value used for imputation, either a numeric or one of "min", "half_min", "small_random",
see above}

\item{na_limit}{only impute features with the proportion of NAs over this limit. For example, if
\code{na_limit = 0.5}, only features with at least half of the values missing are imputed.}
}
\description{
Impute missing values using a simple imputation strategy. All missing values
of a feature are imputed with the same value. It is possible
to only impute features with a large number of missing values this way. This can be useful
for using this function before random forest imputation to speed things up.
The imputation strategies available are:
}
\details{
\itemize{
\item a numeric value: impute all missing values in all features with the same value, e.g. 1
\item "min": impute missing values of a feature with the minimum observed value of that feature
\item "half_min": impute missing values of a feature with half the minimum observed value of that feature
\item "small_random": impute missing values of a feature with random numbers between 0 and the
minimum of that feature (uniform distribution, remember to set the seed number!).
}
}
\examples{
missing <- mark_nas(merged_sample, 0)
imputed <- impute_simple(missing, value = "min")

}
