% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{inspect_dc}
\alias{inspect_dc}
\title{Flag the results of drift correction}
\usage{
inspect_dc(orig, dc, condition = "RSD_r < 0 & D_ratio_r < 0")
}
\arguments{
\item{orig}{a MetaboSet object, before drift correction}

\item{dc}{a MetaboSet object, after drift correction}

\item{condition}{a character specifying the condition, see Details}
}
\value{
MeatboSet object
}
\description{
Chooses whether the drift correction worked well enough by applying
the specified codition for each feature.
If the condition is fulfilled, the drift corrected feature is retained,
otherwise the original feature is retained and the drift corrrected feature is discarded.
The result of this operation is recorded in the feature data
}
\details{
The \code{condition} parameter should be a character giving a condition combatible
with dplyr::filter. The condition is applied on the \strong{changes} in the quality metrics
RSD, RSD_r, D_ratio and D_ratio_r. For example, the default is "RSD_r < 0 and D_ratio_r < 0",
meaning that both RSD_r and D_ratio_r need to decrease in the drift correction, otherwise the
drift corrected feature is discarded and the original is retained.
}
\seealso{
\code{\link{correct_drift}}, \code{\link{save_dc_plots}}
}
