% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install dependencies}
\usage{
install_dependencies(
  preprocessing = TRUE,
  extra = FALSE,
  batch_corr = FALSE,
  misc = FALSE,
  ...
)
}
\arguments{
\item{preprocessing}{logical, install core preprocessing and visualization packages?}

\item{extra}{logical, install extra packages needed for special visualizations and stats?}

\item{batch_corr}{logical, install packages needed for batch effect correction methods?}

\item{misc}{logicl, install miscallenous packages needed for running tests, modifying vignettes etc.?}

\item{...}{other parameters passed to installing functions, like lib}
}
\description{
Attempt to install dependencies package by package, skipping packages with errors
By default, only installs core packages needed for preprocessing. Other packages can
be installed if needed (and the preprocessing packages can be ignored)
}
