% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_normalize}
\alias{inverse_normalize}
\title{Inverse-rank normalization}
\usage{
inverse_normalize(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
MetaboSet object as the one supplied, with normalized features
}
\description{
Applies inverse rank normalization to all features to approximate
a normal distribution.
}
\examples{
normalized <- inverse_normalize(merged_sample)

}
