% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{join_fData}
\alias{join_fData}
\title{Join new columns to feature data}
\usage{
join_fData(object, dframe)
}
\arguments{
\item{object}{a MetaboSet object}

\item{dframe}{a dataframe with the new information}
}
\value{
a MetaboSet object with the new information added to fData(object)
}
\description{
Join a new data frame of information to feature data of a MetaboSet object. The data frame needs to have a column "Feature_ID".
This function is usually used internally by some of the functions in the package, but can be useful.
}
\examples{
new_info <- data.frame(Feature_ID = featureNames(example_set), Feature_number = seq_len(nrow(example_set)))
with_new_info <- join_fData(example_set, new_info)
colnames(fData(with_new_info))

}
