% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_state}
\alias{log_state}
\title{Get the current logging state}
\usage{
log_state()
}
\description{
If a log file is currently in use, prints the log file,
else tells you that logging is not enabled
}
\seealso{
\code{\link{init_log}}, \code{\link{log_text}}, \code{\link{finish_log}}
}
