% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{mark_nas}
\alias{mark_nas}
\title{Mark specified values as missing}
\usage{
mark_nas(object, value)
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{the value to be converted to NA}
}
\value{
MetaboSet object as the one supplied, with missing values correctly set to NA
}
\description{
Replaces all values in the exprs part that equal the specified value with NA.
For example, vendor software might use 0 or 1 to signal a missing value,
which is not understood by R.
}
\examples{
nas_marked <- mark_nas(merged_sample, value = 0)

}
