% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{merge_exprs}
\alias{merge_exprs}
\title{Partially replace exprs field with new values}
\usage{
merge_exprs(object, y)
}
\arguments{
\item{object}{a MetaboSet object}

\item{y}{matrix containing new values to be merged into exprs}
}
\description{
Replaces a subset of data in exprs of an object by new values.
Used after an operation such as imputation is computed only for a subset of features or samples
such as only good-quality features that have not been flagged.
This function is mainly used internally, but can be useful.
}
