% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merge_metabosets}
\alias{merge_metabosets}
\title{Merge MetaboSet objects together}
\usage{
merge_metabosets(..., merge = c("features", "samples"))
}
\arguments{
\item{...}{MetaboSet objects or a list of Metaboset objects}

\item{merge}{what to merge? features is used for combining analytical modes,
samples is used for batches}
}
\value{
A merged MetaboSet object
}
\description{
Merges two or more MetaboSet objects together. Can be used to merge analytical modes or batches.
}
\details{
When merging samples, sample IDs that beging with "QC" or "Ref" are combined so that they have
running numbers on them. This means that if both bathces have samples called "QC_1",
this will not result in an error,
but the sample IDs will be adjusted so that they are unique
}
\examples{
# Merge analytical modes
merged <- merge_metabosets(
  hilic_neg_sample, hilic_pos_sample,
  rp_neg_sample, rp_pos_sample
)
# Merge batches
batch1 <- merged_sample[, merged_sample$Batch == 1]
batch2 <- merged_sample[, merged_sample$Batch == 2]
merged <- merge_batches(batch1, batch2)

}
