% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_sample}
\alias{merged_sample}
\alias{hilic_neg_sample}
\alias{hilic_pos_sample}
\alias{rp_neg_sample}
\alias{rp_pos_sample}
\title{Sample dataset}
\format{
\describe{
  \item{hilic_neg_sample}{features from HILIC column with negative ionization}
  \item{hilic_pos_sample}{features from HILIC column with positive ionization}
  \item{rp_neg_sample}{features from RP column with negative ionization}
  \item{rp_pos_sample}{features from RP column with positive ionization}
  \item{merged_sample}{all features merged together}
}

An object of class \code{MetaboSet} with 21 rows and 218 columns.

An object of class \code{MetaboSet} with 21 rows and 218 columns.

An object of class \code{MetaboSet} with 20 rows and 218 columns.

An object of class \code{MetaboSet} with 21 rows and 218 columns.
}
\usage{
merged_sample

hilic_neg_sample

hilic_pos_sample

rp_neg_sample

rp_pos_sample
}
\description{
Sample data from samples in two groups, at two timepoints.
Contains 83 features from four analytical modes. The analytical modes are
available as separate MetaboSet objects and also as a merged object.
}
\keyword{datasets}
