% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{mixomics_plsda}
\alias{mixomics_plsda}
\title{PLS-DA}
\usage{
mixomics_plsda(object, y, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{y}{character, column name of the grouping variable to predict}

\item{...}{any parameters passed to \code{mixOmics::plsda}}
}
\value{
an object of class "plsdsa"
}
\description{
A simple wrapper for fitting a PLS-DA model using plsda function of the mixOmics package.
The object can then be passed to many of the mixOmics functions for prediction,
performance evaluation etc. To run initial optimization for a PLS-DA model,
try \code{\link{mixomics_plsda_optimize}}.
}
\examples{
plsda_res <- mixomics_plsda(merged_sample, y = "Group")

}
