% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{mixomics_plsda_optimize}
\alias{mixomics_plsda_optimize}
\title{PLS-DA}
\usage{
mixomics_plsda_optimize(object, y, ncomp_max, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{y}{character, column name of the grouping variable to predict}

\item{ncomp_max}{numeric, the maximum number of components to try}

\item{...}{any parameters passed to \code{mixOmics::plsda}}
}
\value{
an object of class "plsdsa"
}
\description{
A wrapper for fitting a PLS-DA model using plsda function of the mixOmics package.
Automatically evaluates performance with different number of components and
chooses the optimal number of components.
}
