% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{mixomics_splsda_optimize}
\alias{mixomics_splsda_optimize}
\title{sPLS-DA}
\usage{
mixomics_splsda_optimize(object, y, ncomp_max, dist, n_features = c(1:10,
  seq(20, 300, 10)), ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{y}{character, column name of the grouping variable to predict}

\item{ncomp_max}{numeric, the maximum number of components to try}

\item{dist}{the distance metric to use, one of "max.dist", "mahalanobis.dist", "centroids.dist"}

\item{n_features}{the number of features to try for each component}

\item{...}{any parameters passed to \code{mixOmics::plsda}}
}
\value{
an object of class "splsdsa"
}
\description{
A wrapper for fitting an sPLS-DA model using splsda function of the mixOmics package.
Automatically evaluates performance with different number of components and
different number of features per component, then chooses the optimal number of components and
optimal number of features for each component.
}
\examples{
set.seed(38)
plsda_res <- mixomics_splsda_optimize(merged_sample, dist = "max.dist", y = "Group", ncomp_max = 2)

}
