% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{mz_rt_plot}
\alias{mz_rt_plot}
\title{Plot m/z vs retention time plot}
\usage{
mz_rt_plot(
  object,
  p_col = NULL,
  mz_col = NULL,
  rt_col = NULL,
  color = NULL,
  color_scale = getOption("notame.color_scale_dis")
)
}
\arguments{
\item{object}{a MetaboSet object or a data frame. If x is a MetaboSet object, fData(x) is used.
If x is a data frame, it is used as is.}

\item{p_col}{the column name containing p-values. This is used to scale the size of the points.}

\item{mz_col, rt_col}{the column names for m/z and retention time. if NULL, automatic detection is attempted.}

\item{color}{the column name used to color the points}

\item{color_scale}{color scale as returned by a ggplot function. Defaults to current continuous color scale.}
}
\value{
a ggplot object
}
\description{
Plots a scatter plot of results of statistical tests, where each point represents a feature.
The plot has retention time on x-axis, m/z on y-axis and the size of the points is scaled based on p-value
}
\examples{


}
