% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunius_correction.R
\name{normalize_batches_brunius}
\alias{normalize_batches_brunius}
\title{Normalize batches}
\usage{
normalize_batches_brunius(object, batch, group = group_col(object),
  ref_label, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch, group}{character, column names of pData with batch labels and group labels}

\item{ref_label}{the label of the reference group i.e. the group that is constant through batches}
}
\value{
list, the object with normalized features and information on which features were corrected by ref samples in each batch.
}
\description{
Normalize bathces by either reference samples of population median.
Uses normalizeBatches function from batchCorr package by Carl Brunius.
}
