% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_auc}
\alias{perform_auc}
\title{Area under curve}
\usage{
perform_auc(
  object,
  time = time_col(object),
  subject = subject_col(object),
  group = group_col(object)
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{time, subject, group}{column names of pData(object), holding time, subject and group labels}
}
\value{
a pseudo MetaboSet object with the AUCs
}
\description{
Compute area under curve (AUC) for each subject and feature.
Creates a pseudo MetaboSet object, where the "samples" are subjects
(or subject/group combinations in case the same subjects are submitted to different treatments)
and the "abundances" are AUCs. This object can then be used to compute results of e.g. t-tests of
AUCs between groups.
}
\examples{
# Drop QC samples before computing AUCs
aucs <- perform_auc(drop_qcs(example_set))
# t-test with the AUCs
t_test_results <- perform_t_test(aucs, formula_char =  "Feature ~ Group")

}
\seealso{
\code{\link[PK]{auc}}
}
