% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_conover_test}
\alias{perform_conover_test}
\title{Conover-Iman test}
\usage{
perform_conover_test(
  object,
  formula_char,
  all_features = FALSE,
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{formula_char}{character, the formula to be used in the linear model (see Details)}

\item{all_features}{should all features be included in FDR correction?}

\item{separate_by}{character, name of the column on which you want to separate the tests.
Usually used when dealing with multiple sample types in the same object.}
}
\value{
data frame containing Conover-Iman t-test statistics, p-values and FDR corrected p-values of each comparison
for each feature
}
\description{
Wrapper for \code{\link[PMCMRplus]{kwAllPairsConoverTest}}.
Performs Conover's non-parametric all-pairs comparison test for
Kruskal-type ranked data. Should only be used as a post-hoc test after
a significant Kruskal-Wallis test.
}
\details{
Argument \code{p.adjust.method} is hard-coded to "none" as FDR correction is
done outside \code{kwAllPairsConoverTest} similarly to other tests in notame.
}
\examples{
# Basic example
conover_res <- perform_conover_test(example_set, "Feature ~ Group")
# Separately, usually done when dealing with multiple sample types in the same object
conover_res <- perform_conover_test(drop_qcs(example_set), "Feature ~ Time", separate_by = "Group")
}
\seealso{
\code{\link[PMCMRplus]{kwAllPairsConoverTest}}
}
