% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_correlation_tests}
\alias{perform_correlation_tests}
\title{Perform correlation tests}
\usage{
perform_correlation_tests(
  object,
  x,
  y = x,
  object2 = NULL,
  fdr = TRUE,
  duplicates = FALSE,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{x}{character vector, names of variables to be correlated}

\item{y}{character vector, either identical to x (the default) or a distinct set of variables
to be correlated agains x}

\item{object2}{optional second MeatboSet object. If provided, x variables will be taken from object and
y variables will be taken from object2. Both objects should have the same number of samples.}

\item{fdr}{logical, whether p-values from the correlation test should be adjusted with FDR correction}

\item{duplicates}{logical, whether correlations should be dublicated. If \code{TRUE}, each correlation
will be included in the results twice, where the order of the variables (which is x and which is y)
is changed. Can be useful for e.g. plotting a heatmap of the results, see examples of
\code{\link{plot_effect_heatmap}}}

\item{...}{other parameters passed to \code{\link{cor.test}}, such as method}
}
\value{
a data frame with the results of correlation tests: the pair of variables, correlation coefficient
and p-value
}
\description{
Performs a correlation test between two sets of variables. All the variables must be either
feature names or column names of pheno data (sample information). There are two ways to use this function:
either provide a set of variables as \code{x}, and all correlations between those variables are computed. Or
provide two distinct sets of variables \code{x, y} and correlations between each x variable
and each y variable are computed.
}
\examples{
# Correlations between all features
correlations <- perform_correlation_tests(example_set, x = featureNames(example_set))

# Spearman Correlations between features and sample information variables
# Drop QCs and convert time to numeric
no_qc <- drop_qcs(example_set)
no_qc$Time <- as.numeric(no_qc$Time)
correlations <- perform_correlation_tests(no_qc, x = featureNames(example_set),
                                         y = c("Time", "Injection_order"), method = "spearman")

# Correlations between variables from two distinct MetaboSets
cross_object_cor <- perform_correlation_tests(hilic_neg_sample, x = featureNames(hilic_neg_sample),
                                              object2 = hilic_pos_sample, y = featureNames(hilic_pos_sample))
}
\seealso{
\code{\link{cor.test}}
}
