% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_friedman_test}
\alias{perform_friedman_test}
\title{Friedman test}
\usage{
perform_friedman_test(
  object,
  group,
  id,
  all_features = FALSE,
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the column containing the grouping variable}

\item{id}{character, name of the column containing the pairing variable.}

\item{all_features}{should all features be included in FDR correction?}

\item{separate_by}{character, name of the column on which you want to separate the tests. Usually used when dealing}
}
\value{
data frame containing Friedman test statistics and p-values
for each feature
}
\description{
Wrapper for the \code{\link[PMCMRplus]{friedmanTest}}. Performs Friedman test
for a two-way balanced complete block design.
}
\examples{
# Basic example
friedman_res <- perform_friedman_test(drop_qcs(example_set), group = "Time", id = "Subject_ID")

}
\seealso{
\code{\link[PMCMRplus]{friedmanTest}}
}
