% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_paired_conover_test}
\alias{perform_paired_conover_test}
\title{Paired Conover-Iman test}
\usage{
perform_paired_conover_test(
  object,
  group,
  id,
  all_features = FALSE,
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the column containing the grouping variable}

\item{id}{character, name of the column containing the pairing variable.}

\item{all_features}{should all features be included in FDR correction?}

\item{separate_by}{character, name of the column on which you want to separate the tests. Usually used when dealing}
}
\value{
data frame containing Conover-Iman t-test statistics, p-values and FDR corrected p-values of each comparison
for each feature
}
\description{
Wrapper for the \code{\link[PMCMRplus]{frdAllPairsConoverTest}}. Performs
pairwise comparisons using Conover's all-pairs test for a two-way balanced
complete block design. Should only be used as a post-hoc test after a
significant Friedman test.
}
\details{
Argument \code{p.adjust.method} is hard-coded to "none" as FDR correction is done outside
\code{kwAllPairsConoverTest} similarly to other tests in notame.
}
\examples{
# Basic example
conover_res <- perform_paired_conover_test(drop_qcs(example_set), group = "Time", id = "Subject_ID")
}
\seealso{
\code{\link[PMCMRplus]{frdAllPairsConoverTest}}
}
