% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_paired_t_test}
\alias{perform_paired_t_test}
\title{Perform paired t-tests}
\usage{
perform_paired_t_test(object, group, id, all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, column name of pData with the group information}

\item{id}{character, column name of pData with the identifiers for the pairs}

\item{all_features}{should all features be included in FDR correction?}

\item{...}{additional parameters to t.test}
}
\value{
data frame with the results
}
\description{
Performs paired t-tests between two groups.
}
\examples{
paired_t_results <- perform_paired_t_test(drop_qcs(example_set), group = "Time", id = "Subject_ID")

}
\seealso{
\code{\link{t.test}}
}
