% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_pairwise_non_parametric}
\alias{perform_pairwise_non_parametric}
\title{Perform pairwise  non-parametric tests}
\usage{
perform_pairwise_non_parametric(
  object,
  group = group_col(object),
  is_paired = FALSE,
  id = NULL,
  all_features = FALSE,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, column name of phenoData giving the groups}

\item{is_paired}{logical, use pairwise tests}

\item{id}{character, name of the subject identification column for paired version}

\item{all_features}{should all features be included in FDR correction?}

\item{...}{other parameters passed to test functions}
}
\value{
data frame with the results
}
\description{
Performs pairwise non-parametric tests between all study groups.
Use \code{is_paired = FALSE} for Mann-Whitney u-tests
Use \code{is_paired = TRUE} for Wilcoxon signed rank tests.
NOTE! Does not use formula interface
}
\details{
P-values of each comparison are corrected separately from each other.
}
\examples{
# Including QCs as a study group for example
mann_whitney_results <- perform_pairwise_non_parametric(merged_sample, group = "Group")
# Using paired mode (pairs with QC are skipped as there are no common IDs in 'example_set')
wilcoxon_signed_results <- perform_pairwise_non_parametric(example_set, group = "Time", is_paired = TRUE, id = "Subject_ID")

}
\seealso{
\code{\link{perform_mann_whitney}},
\code{\link{perform_wilcoxon_signed_rank}},
\code{\link{wilcox.test}}
}
