% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{perform_permanova}
\alias{perform_permanova}
\title{Perform PERMANOVA}
\usage{
perform_permanova(object, all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{should all features be included in FDR correction?}

\item{...}{other parameters to \code{\link{PERMANOVA}}}

\item{transform}{Transformation to use in \code{IniTransform}}

\item{coef}{Coefficient to calculate continuous distances in \code{DistContinuous}. By default we use Pythagorean distances.}
}
\value{
PERMANOVA object
}
\description{
Performs permutational multivariate analysis of variance.
}
