% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{perform_repeatability}
\alias{perform_repeatability}
\title{Compute repeatability measures}
\usage{
perform_repeatability(object, group)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{column name of pData givinh the group labels}
}
\value{
A data frame with one row per feature with the repeatability measure.
}
\description{
Computes repeatability for each feature with the following formula:
\deqn{\frac{\sigma^2_{between}}{\sigma^2_{between} + \sigma^2_{within}}}
The repeatability ranges from 0 to 1. Higher repeatability depicts less
variation between batches.
}
\examples{
# Batch correction
replicates <- list(which(merged_sample$QC == "QC"))
batch_corrected <- ruvs_qc(merged_sample, 
  batch = "Batch", replicates = replicates)
# Evaluate batch correction
rep_orig <- perform_repeatability(merged_sample, group = "Group")
mean(rep_orig$Repeatability, na.rm = TRUE)
rep_corr <- perform_repeatability(batch_corrected, group = "Group")
mean(rep_corr$Repeatability, na.rm = TRUE)

}
