% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_wilcoxon_signed_rank}
\alias{perform_wilcoxon_signed_rank}
\title{Perform Wilcoxon signed rank test}
\usage{
perform_wilcoxon_signed_rank(object, group, id, all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the group column}

\item{id}{character, column name of pData with the identifiers for the pairs}

\item{all_features}{should all features be included in FDR correction?}

\item{...}{other parameters to \code{\link{wilcox.test}}}
}
\value{
data frame with the results
}
\description{
Performs Wilcoxon signed rank test
Uses base R function \code{wilcox.test} with \code{paired = TRUE}.
}
\examples{
perform_wilcoxon_signed_rank(drop_qcs(example_set), group = "Time", id = "Subject_ID")

}
\seealso{
\code{\link{wilcox.test}}
}
