% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_injection_lm}
\alias{plot_injection_lm}
\title{Estimate the magnitude of drift}
\usage{
plot_injection_lm(object)
}
\arguments{
\item{object}{A MetaboSet object}
}
\value{
A ggplot object
}
\description{
Plots histograms of p-values from linear regression model, where each feature is predicted
by injection order alone. The expected uniform distribution is represented by a dashed red line.
}
\seealso{
\code{\link{plot_p_histogram}}
}
