% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_p_histogram}
\alias{plot_p_histogram}
\title{Histogram of p-values}
\usage{
plot_p_histogram(p_values)
}
\arguments{
\item{p_values}{list, each element is a vector of p-values. The list names are used as plot titles}
}
\value{
A ggplot object.
}
\description{
Draws histograms of p-values with expected uniform distribution represented by a dashed red line
}
