% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA plot}
\usage{
plot_pca(object, center = TRUE, scale = "uv", method = "ppca",
  color = group_col(object), shape = NULL, density = TRUE,
  title = "PCA", subtitle = NULL, color_scale = NULL,
  shape_scale = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{center}{logical, should the data  be centered prior to PCA (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{method}{the method to use, see documentation in pcaMethods}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{density}{logical, whether to include density plots to both axes}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{shape_scale}{the shape scale as returned by a ggplot function}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
if
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods.
}
\seealso{
\code{\link[pcaMethods]{pca}}
}
