% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca_loadings}
\alias{plot_pca_loadings}
\title{PCA loadings plot}
\usage{
plot_pca_loadings(object, all_features = FALSE, center = TRUE,
  scale = "uv", npc1 = 10, npc2 = 10, title = "PCA loadings",
  subtitle = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{center}{logical, should the data be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{npc1, npc2}{number of top feature to plot for first and second principal component}

\item{title, subtitle}{the titles of the plot}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
a ggplot object.
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods and plots the loadings of first principal components
\strong{CITATION:} When using this function, cite the \code{pcaMethods} package
}
\seealso{
\code{\link[pcaMethods]{pca}}
}
