% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plot_pls}
\alias{plot_pls}
\title{Plot points in PLS space}
\usage{
plot_pls(model, Y, y, title)
}
\arguments{
\item{model}{a PLS or sPLS model}

\item{Y}{the Y matrix}

\item{y}{the name of the y variable}

\item{title}{plot title}
}
\description{
A helper function for \code{mixomics_pls} and \code{mixomics_spls}.
}
