% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Draw heatmaps}
\usage{
plot_sample_heatmap(
  object,
  all_features = FALSE,
  dist_method = "euclidean",
  clust_method = "ward.D2",
  center = TRUE,
  scale = "uv",
  group_bar = TRUE,
  group = group_col(object),
  title = "Heatmap of distances between samples",
  subtitle = NULL,
  fill_scale_con = getOption("notame.fill_scale_con"),
  fill_scale_dis = getOption("notame.fill_scale_dis")
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{dist_method}{distance method used in clustering, see \code{\link{dist}}}

\item{clust_method}{clustering method used in clustering, see \code{\link{hclust}}}

\item{center}{logical, should the data  be centered?}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{group_bar}{logical, should a bar showing the groups be drawn under the heat map?}

\item{group}{character, name of the column used for coloring the group bar}

\item{title}{The plot title}

\item{subtitle}{The plot subtitle}

\item{fill_scale_con}{Continuous fill scale for the heatmap as returned by a ggplot function}

\item{fill_scale_dis}{Discrete fill scale for the group bar as returned by a ggplot function}
}
\value{
a ggplot object. If \code{group_bar} is \code{TRUE}, the plot will consist of multiple
parts and is harder to modify.
}
\description{
Draws a heatmap of the distances between the samples of an experiment,
the samples are ordered by hierarchical clustering.
}
\examples{
plot_sample_heatmap(merged_sample)

}
\seealso{
\code{\link{dist}} \code{\link{hclust}}
}
