% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suprahex_plots.R
\name{plot_sample_suprahex}
\alias{plot_sample_suprahex}
\title{Suprahex plots}
\usage{
plot_sample_suprahex(object, file, width = 16, height = 16,
  all_features = FALSE, sample_labels = "Sample_ID",
  grid_xdim = NULL, grid_ydim = NULL, title.xy = c(0.35, 1),
  title.rotate = 0, fontsize = 10, colormap = "jet", ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{file}{filename (pdf)}

\item{width, height}{dimensions of the plot}

\item{all_features}{if FALSE, flagged features are droppped}

\item{sample_labels}{the column for labels of samples in the plot}

\item{grid_xdim, grid_ydim}{dimensions of the grid for the samples}

\item{title.xy}{position of the sample label relative to the supraHex}

\item{title.rotate}{rotation of the sample label in degrees}

\item{fontsize}{the fontsize for sample labels}

\item{colormap}{colormap for the hexagons}

\item{...}{other parameters for supraHex::sPipeline}
}
\description{
Plots supraHex plots of each sample using functions from the supraHex package.
See the supraHex paper and package vignette for more information.
}
\examples{
plot_sample_suprahex(merged_sample[, 1:20], file = "supra.pdf", xdim = 5, title.xy = c(0.35, 1),
                     width = 10, height = 10, grid_xdim = 7, grid_ydim = 7, sample_labels = "Group")

}
\seealso{
\code{\link[supraHex]{sPipeline}}, \code{\link[supraHex]{sCompReorder}}, \code{\link[supraHex]{visCompReorder}}
}
