% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_tsne}
\alias{plot_tsne}
\title{t-SNE scatter plot}
\usage{
plot_tsne(object, center = TRUE, scale = "uv", perplexity = 30,
  color = group_col(object), shape = NULL, density = FALSE,
  title = "t-SNE", subtitle = paste("Perplexity:", perplexity),
  color_scale = NULL, shape_scale = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{center}{logical, should the data  be centered prior to PCA (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{perplexity}{the perplexity used in t-SNE}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{density}{logical, whether to include density plots to both axes}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{shape_scale}{the shape scale as returned by a ggplot function}

\item{...}{additional arguments passed to Rtsne::Rtsne}
}
\value{
if \code{density} is \code{TRUE}, an object of class gTable (from the cowplot package).
NOTE this object must be drawn with ggdraw() instead of plot(). Otherwise, a ggplot object.
}
\description{
Computes t-SNE into two dimensions and plots the map points.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
