% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_tsne_hexbin}
\alias{plot_tsne_hexbin}
\title{t-SNE hexbin plot}
\usage{
plot_tsne_hexbin(object, all_features = FALSE, center = TRUE,
  scale = "uv", perplexity = 30, fill = "Injection_order",
  summary_fun = "mean", bins = 10, title = "t-SNE",
  subtitle = paste("Perplexity:", perplexity), fill_scale = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{center}{logical, should the data  be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{perplexity}{the perplexity used in t-SNE}

\item{fill}{character, name of the column used for coloring the hexagons}

\item{summary_fun}{the function used to compute the value for each hexagon}

\item{bins}{the number of bins in x and y axes}

\item{fill_scale}{the fill scale as returned by a ggplot function}

\item{...}{additional arguments passed to Rtsne::Rtsne}

\item{color}{character, name of the column used for coloring the points}
}
\value{
A ggplot object.
}
\description{
Computes t-SNE into two dimensions and plots the map points.
\strong{CITATION:} When using this function, cite the \code{pcaMethods} and \code{Rtsne} packages
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
