% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{pqn_normalization}
\alias{pqn_normalization}
\title{Probabilistic quotient normalization}
\usage{
pqn_normalization(
  object,
  ref = c("qc", "all"),
  method = c("median", "mean"),
  all_features = FALSE
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{ref}{character, the type of reference samples to use for normalization.}

\item{method}{character, the method to use for calculating the reference sample.}

\item{all_features}{logical, should all features be used for calculating the reference sample?}
}
\value{
a MetaboSet object with altered feature abundances
}
\description{
Apply probabilistic quotient normalization (PQN) to the exprs part of a MetaboSet object. By default, reference
is calculated from high-quality QC samples and the median of the reference is used for normalization.
Check parameters for more options.
}
\examples{
pqn_set <- pqn_normalization(merged_sample)

}
