% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_found}
\alias{prop_found}
\title{Proportion of non-missing values in a vector}
\usage{
prop_found(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric, the proportion of non-missing values in vector.
}
\description{
Proportion of non-missing values in a vector
}
\examples{
example_set <- mark_nas(example_set, value = 0)
prop_found(exprs(example_set))

}
