% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_na}
\alias{prop_na}
\title{Proportion of NA values in a vector}
\usage{
prop_na(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric, the proportion of non-missing values in a vector.
}
\description{
Proportion of NA values in a vector
}
\examples{
example_set <- mark_nas(example_set)
prop_na(exprs(example_set))

}
