% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{pull_clusters}
\alias{pull_clusters}
\title{Extract information of the features in clusters}
\usage{
pull_clusters(data, features, name_col)
}
\arguments{
\item{data}{data frame of the original LC-MS data}

\item{features}{data frame holding the feature information}

\item{name_col}{name_col character, name of the column in features that
contains feature names}
}
\value{
A list of two items:
\itemize{
\item cdata: a new data frame with the combined LC-MS data
\item cfeatures: data frame, feature information per cluster
}
}
\description{
For each cluster, the LC-MS data of the feature with largest median peak 
area is retained, all the features inside every cluster are recorded.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
data <- combined_data(example_set)
features <- fData(example_set)
features$MPA <- sapply(data[, features[, "Feature_ID"]], finite_median)
conn <- find_connections(data = data, features = features,
  corr_thresh = 0.4, rt_window = 2,
  name_col = "Feature_ID", mz_col = "Mass", rt_col = "RetentionTime")
clusters <- find_clusters(connections = conn, d_thresh = 0.6)
features_clustered <- assign_cluster_id(data, clusters, features,
  name_col = "Feature_ID")
visualize_clusters(data, features, clusters, min_size = 3,
  rt_window = 2, name_col = "Feature_ID", mz_col = "Mass",
  rt_col = "RetentionTime", file_path = "./clusters")
pulled <- pull_clusters(data, features_clustered, name_col = "Feature_ID")
\dontshow{setwd(.old_wd)}

}
