% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{pull_features}
\alias{pull_features}
\title{Extract information of the features in clusters}
\usage{
pull_features(clusters, data, features, name_col)
}
\arguments{
\item{clusters}{list of cluster information, as returned by find_clusters}

\item{data}{data frame of the original LC-MS data}

\item{features}{data frame holding the feature information}

\item{name_col}{name_col character, name of the column in features that contains feature names}
}
\value{
a list of two items:
\begin{itemize}
\item cdata: a new data frame with the combined LC-MS data
\item cfeatures: data frame, feature information per cluster
\end{itemize}
}
\description{
The LC-MS data of the feature with largest median peak area is retained,
all the features in every cluster are recorded
}
