% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rank_biserial_correlation}
\alias{rank_biserial_correlation}
\title{Rank-biserial correlation}
\usage{
rank_biserial_correlation(
  object,
  group = group_col(object),
  id = NULL,
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the group column}

\item{id}{character, name of the column containing subject IDs}

\item{separate_by}{character, name of the column on which you want to
separate the tests. Usually used when dealing with multiple sample types
in the same object.}
}
\description{
Computes Glass's rank-biserial correlation for each feature. Treats the first level of grouping variable
as the reference group. Therefore, the difference in rank means is calculated as (second level - first level).
If \code{id} is specified, the correlation is computed for each pair of samples with the same ID.
}
\details{
Correlation coefficient can be interpreted as proportion of favourable outcomes in the second group.
Here favourable means that the ranks are lower in the second group and therefore feature values are higher.
Thus, negative values mean that the first group have higher feature values.
}
\examples{
# Basic example
rrc <- rank_biserial_correlation(drop_qcs(example_set), group = "Group")
# Paired
rrc_paired <- rank_biserial_correlation(drop_qcs(example_set), group = "Time", id = "Subject_ID")
# Separately to each time point in each group
rrc_within_groups <- rank_biserial_correlation(drop_qcs(example_set), group = "Time", separate_by = "Group")
}
