% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{read_from_excel}
\alias{read_from_excel}
\title{Read formatted Excel files}
\usage{
read_from_excel(file, sheet, corner_row, corner_column,
  id_prefix = "ID_", split_by = NULL, name = NULL)
}
\description{
Reads data from an Excel file of the following format:
\itemize {
  \item Left side of the sheet contains information about the features, size features x feature info columns
  \item Top part contains sample information, size sample info variables x samples
  \item The middle contains the actual abundances, size features x samples
}
See the vignette for more information.
This function separates the three parts from the file, and returns them in a list
}
\details{
@param file path to the Excel file
 @param sheet the sheet number or name
 @param corner_row interger, the bottom row of sample information, usually contains data file names and feature info column names
 @param corner_column integer or character, the corresponding column number or the column name (letter) in Excel
 @param id_prefix character, prefix for  autogenerated sample IDs, see Details
 @param split_by character vector, in the case where all the modes are in the same Excel file,
 the column names of feature data used to separate the modes (usually Mode and Column)
 @param name in the case where the Excel file only contains one mode, the name of the mode, such as "Hilic_neg"

@return list of three data frames:
\itemize {
  \item assay_data: the actual abundances, size features x samples
  \item pheno_data: sample information, size sample info variables x samples
  \item feature_data: information about the features, size features x feature info columns
}

@details Only specify one of \code{split_by} and \code{name}. The feature data returned will contain a column
 named "Split", which is used to separate features from different modes. Unless a column named "Feature_ID"
 is found in the file, a feature ID will be generated based on the value of "Split", mass and retention time.
 The function will try to find columns for mass and retention time by looking at a few common alternatives,
 and throw an error if no matching column is found.
}
