% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{save_batch_plots}
\alias{save_batch_plots}
\title{Save batch correction plots}
\usage{
save_batch_plots(orig, corrected, file, batch = "Batch",
  color = "Batch", shape = "QC", color_scale = NULL,
  shape_scale = NULL)
}
\arguments{
\item{orig, corrected}{MetaboSet objects before and after batch effect correction}

\item{file}{path to the PDF file where the plots will be saved}

\item{batch, color, shape}{column names of pData for batch labels,
and column used for coloring and shaping points (by default batch and QC)}

\item{color_scale, shape_scale}{scales for color and scale as returned by ggplot functions.}
}
\description{
Saves plots of each feature showing the effect of batch correction.
Plots show QC samples and regular samples inside each batch, plus the
batch mean for all samples and QC samples as a horizontal line.
NOTE: if you change the shape variable, be sure to set a shape scale as well,
the default scale only has 2 values, so it can only accomodate 2 shapes.
}
