% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{save_dc_plots}
\alias{save_dc_plots}
\title{Drift correction plots}
\usage{
save_dc_plots(orig, dc, file, width = 8, height = 6,
  color = group_col(orig), shape = NULL, color_scale = NULL)
}
\arguments{
\item{orig}{a MetaboSet object, before drift correction}

\item{dc}{a MetaboSet object, after drift correction as returned by correct_drift}

\item{file}{path to the PDF file where the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\description{
Plots the data before and after drift correction, with the regression line drawn with
the original data.
}
\details{
If \code{shape} is set to \code{NULL} (the default), the column used for color
is also used for shape
}
\seealso{
\code{\link{correct_drift}}, \code{\link{inspect_dc}}
}
