% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_feature_plots}
\alias{save_feature_plots}
\title{Save plots of individual features}
\usage{
save_feature_plots(
  object,
  file_path,
  format,
  title,
  subtitle,
  text_base_size,
  plot_fun,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{file_path}{character, a file path for PDF or prefix added to the file paths for other formats}

\item{format}{character, format in which the plots should be saved}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle}

\item{text_base_size}{integer, base size for text in figures}

\item{plot_fun}{a function with arguments:
data frame from combined_data(object)
feature id
Should return a ggplot object for plotting}

\item{...}{other arguments to plotting function}
}
\description{
Helper function for saving plots of individual features
to either one multi-page PDF or separate EMF figures
}
