% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_group_lineplots}
\alias{save_group_lineplots}
\title{Save line plots with errorbars by group}
\usage{
save_group_lineplots(
  object,
  all_features = FALSE,
  prefix,
  format,
  width = 8,
  height = 6,
  res = 300,
  x = time_col(object),
  group = group_col(object),
  title = "Feature_ID",
  subtitle = NULL,
  fun.data = "mean_cl_boot",
  fun.y = NULL,
  fun.ymin = NULL,
  fun.ymax = NULL,
  position_dodge_amount = 0.2,
  color_scale = getOption("notame.color_scale_dis")
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{prefix}{character, a file path prefix added to the file paths}

\item{format}{character, format in which the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{res}{integer, resolution in ppi for non-vector formats}

\item{x}{character, name of the column to be used as x-axis}

\item{group}{character, name of the column containing group information, used for coloring}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{fun.data}{passed to ggplot2::stat_summary and used for errorbars,
"A function that is given the complete data and should return a data frame with variables ymin, y, and ymax."}

\item{fun.ymin, fun.y, fun.ymax}{Alternative to fun.data, passed to ggplot2::stat_summary,
"supply three individual functions that are each passed a vector of x's and should return a single number"}

\item{position_dodge_amount}{numeric: how much the group mean points should dodge away from each other}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\description{
Plots the change in the feature abundances as a function of e.g. time.
A line is drawn for each group and error bars are added.
A separate plot is drawn for each feature.
}
\examples{
\dontrun{save_group_lineplots(drop_qcs(merged_sample),
                    prefix = "./group_line_plots/",
                    format = "pdf"
                    )}

}
\seealso{
\code{\link[notame]{save_plot}},
\code{\link[ggplot2]{stat_summary}}
}
