% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{save_lm_diagnostic_plots}
\alias{save_lm_diagnostic_plots}
\title{Plot model diagnostics}
\usage{
save_lm_diagnostic_plots(object, formula_char, model_type = c("lm",
  "lmer"), file, all_features = FALSE, width = 12, height = 18, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{formula_char}{character, the formula to be used in the linear model (see Details)}

\item{model_type}{the type of model, either "lm" or "lmer"}

\item{file}{the file where the plots will be saved}

\item{all_features}{should all features be included?}

\item{width, height}{the dimension of the plot}

\item{...}{other parameters passed to lm/lmer and/or autoplot function from the ggfortify package}
}
\description{
Plot model diagnostics for lm or lmer models fit separately for each feature.
One page of plots is saved per feature. Uses autoplot function from ggfortify
}
\examples{
save_lm_diagnostic_plots(example_set, formula_char = "Feature ~ Group", model_type = "lm", file = "test.pdf")

# Add colouring by time point (don't mind the warnings)
save_lm_diagnostic_plots(example_set, formula_char = "Feature ~ Group",
                         model_type = "lm", file = "test.pdf", colour = "Group")

# Linear mixed model
save_lm_diagnostic_plots(example_set, formula_char = "Feature ~ Group + (1|Subject_ID)",
                         model_type = "lmer", file = "test.pdf")


}
