% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{scale,MetaboSet-method}
\alias{scale,MetaboSet-method}
\title{Scale exprs data}
\usage{
\S4method{scale}{MetaboSet}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a MetaboSet object}

\item{center, scale}{as in base scale function}
}
\description{
Applies the base R function scale to transposed exprs matrix. See ?scale for details
}
