% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{subject_col}
\alias{subject_col}
\alias{subject_col<-}
\title{Get and set the name of special column for subject identifiers}
\usage{
subject_col(object)

subject_col(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{string, name of column to be designated for holding subject 
identifiers}
}
\value{
Character, the name of the subject variable.

For the endomorphism, an object with the subject variable set to the 
specified variable.
}
\description{
Get and set the name of special column for subject identifiers
}
\examples{
# Get name of subject variable
subject_col(example_set)
# Set subject variable
subject_col(example_set) <- "Subject_ID"
}
