% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Summary statistics}
\usage{
summary_statistics(object, grouping_cols = group_col(object))
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
a data frame with the summary statistics
}
\description{
Computes chosen summary statistics for each feature,
possibly grouped by a factor
}
