% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{time_col}
\alias{time_col}
\alias{time_col<-}
\title{These generic functions access and set the special column for time point}
\usage{
time_col(object)

time_col(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{string, name of column to be designated for holding time point}
}
\description{
These generic functions access and set the special column for time point
}
\examples{
# Get name of time variable
time_col(example_set)

# Set time variable
time_col(example_set) <- "Time"

}
