% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize clusters of features}
\usage{
visualize_clusters(
  data,
  features,
  clusters,
  min_size,
  rt_window,
  name_col,
  mz_col,
  rt_col,
  file_path
)
}
\arguments{
\item{data}{data frame with the abundances of features, with features as columns}

\item{features}{data frame with feature information, fData(object)}

\item{clusters}{a list of clusters as returned by find_clusters}

\item{min_size}{the minimum number of features a cluster needs to have to be plotted}

\item{rt_window}{numeric, the retention time window to use in linking features. NOTE you
need to use the same unit as in the retention time column}

\item{name_col}{character, name of the column in features that contains feature names}

\item{mz_col}{character, name of the column in features that contains mass-to-charge ratios}

\item{rt_col}{character, name of the column in features that contains retention times}

\item{file_path}{the prefix to the files to be plotted}
}
\value{
a data frame of pairs of signals that are linked together
\itemize{
\item x & y: indexes and names of the signals
\item cor: correlation coefficient
\item mz_diff & rt_diff: mass and retention time difference
}
}
\description{
Draws multiple visualizations of each cluster, creating a separate file for each cluster.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
data <- combined_data(example_set)
features <- fData(example_set)
features$MPA <- sapply(data[, features[, "Feature_ID"]], finite_median)
conn <- find_connections(data = data, features = features,
  corr_thresh = 0.4, rt_window = 2,
  name_col = "Feature_ID", mz_col = "Mass", rt_col = "RetentionTime")
clusters <- find_clusters(connections = conn, d_thresh = 0.6)
features_clustered <- assign_cluster_id(data, clusters, features,
  name_col = "Feature_ID")
visualize_clusters(data, features, clusters, min_size = 3,
  rt_window = 2, name_col = "Feature_ID", mz_col = "Mass",
  rt_col = "RetentionTime", file_path = "./clusters")
pulled <- pull_clusters(data, features_clustered, name_col = "Feature_ID")
\dontshow{setwd(.old_wd)}

}
