% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  object,
  x,
  p,
  p_fdr = NULL,
  color = NULL,
  p_breaks = c(0.05, 0.01, 0.001, 1e-04),
  fdr_limit = 0.05,
  log2_x = FALSE,
  center_x_axis = TRUE,
  x_lim = NULL,
  color_scale = getOption("notame.color_scale_con"),
  title = "Volcano plot",
  subtitle = NULL,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object or a data frame. If x is a MetaboSet object, fData(x) is used.
If x is a data frame, it is used as is.}

\item{x, p}{the column names of effect size (x-axis) and p-values}

\item{p_fdr}{column name of FDR corrected p-values, used to draw a line showing the fdr-corrected significance level}

\item{color}{column name used to color the plots}

\item{p_breaks}{a numerical vector of the p_values to show on the y-axis}

\item{fdr_limit}{the significance level used in the experiment}

\item{log2_x}{logical, whether effect size should be plotted on a log2 axis}

\item{center_x_axis}{logical, whether x-axis should be centered. If \code{TRUE}, the "zero-effect" will
be on the middle of the plot. The "zero effect" is 0 if \code{log2_x = FALSE} and 1 if  \code{log2_x = TRUE}}

\item{x_lim}{numerical vector of length 2 for manually setting the x-axis limits}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{title, subtitle}{the title and subtitle of the plot}

\item{...}{parameters passed to \code{\link[ggplot2]{geom_point}}, such as shape and alpha values. New aesthetics can
also be passed using \code{mapping = aes(...)}.}
}
\value{
a ggplot object
}
\description{
Draws a volcano plot of effect size and p-values.
}
\examples{
# naturally, this looks messy as there are not enough p-values
lm_results <- perform_lm(drop_qcs(merged_sample), formula_char = "Feature ~ Group")
volcano_plot(lm_results, x = "GroupB_Estimate",
             p = "GroupB_P", p_fdr = "GroupB_P_FDR",
             fdr_limit = 0.1)

}
