% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{align_batches}
\alias{align_batches}
\title{Align features between batches}
\usage{
align_batches(
  object_na,
  object_fill,
  batch,
  mz,
  rt,
  NAhard = 0.8,
  mzdiff = 0.002,
  rtdiff = 15,
  plot_folder = NULL
)
}
\arguments{
\item{object_na}{a MetaboSet object with missing values as NA}

\item{object_fill}{a similar MetaboSet object with imputed values
(used to compute distances between features, can contain missing values as 
well)}

\item{batch}{character, column name of pData with batch labels}

\item{mz, rt}{column names of m/z and retention time columns in fData}

\item{NAhard}{proportion of NAs within batch for feature to be considered
missing}

\item{mzdiff, rtdiff}{the windows for m/z and retention time for aligning 
features}

\item{plot_folder}{path to the location where the plots should be saved, if 
NULL, no plots are saved}
}
\value{
A MetaboSet object with the aligned features.
}
\description{
Aligns features with m/z or retention time shift between batches using
alignBatches from batchCorr package. See more details in the original paper.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
fData(example_set)[3, ]$Average_Mz <- fData(example_set)[2, ]$Average_Mz + 
0.001
# Initialize objects
example_set_fill <- example_set
example_set_na <- example_set
# Introduce over 80\% missing values in QC samples per batch of two features 
# so they are considered "missing", with orthogonal batch presence
exprs(example_set_na)[2, c(1, 7, 13, 19)] <- NA
exprs(example_set_na)[
  3, c(26, 32, 38, 44)] <- NA
batch_aligned <- align_batches(example_set_na, example_set_fill, 
  batch = "Batch", mz = "Average_Mz", rt = "Average_Rt_min", 
  mzdiff = 0.002, rtdiff = 15, plot_folder = "./Figures")
\dontshow{setwd(.old_wd)}

}
\seealso{
\code{\link[batchCorr]{alignBatches}}
}
