% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{assess_quality}
\alias{assess_quality}
\title{Assess quality information of features}
\usage{
assess_quality(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
a MetaboSet object with quality metrics in fData.
}
\description{
Assess quality information of features
}
\examples{
ex_set <- assess_quality(merged_sample)
fData(ex_set)

}
