% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{citations}
\alias{citations}
\title{Show citations}
\usage{
citations()
}
\value{
None, the function is invoked for its side effect.
}
\description{
This function lists citations behind the notame functions that have been 
called during the session. All notame functions 
update the list automatically. The citations are taken from the call to 
'\code{citation("package")}, and complemented with a brief description of 
what the package was used for.
NOTE: the citations might not point to the correct paper if the package 
authors have not supplied correct citation information for their package.
The output is written to the current log file, if specified.
}
\examples{
citations()
data(example_set)
ex_set <- flag_quality(example_set)
# Broadhurst et al.(2018) added to citations
citations()

}
