% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cohens_d}
\alias{cohens_d}
\title{Cohen's D}
\usage{
cohens_d(object, id = subject_col(object), group = group_col(object),
  time = time_col(object))
}
\arguments{
\item{object}{a MetaboSet object}

\item{id}{character, name of the subject ID column}

\item{group}{character, name of the group column}

\item{time}{character, name of the time column}
}
\value{
data frame with Cohen's d for each feature
}
\description{
Computes Cohen's D for each feature
}
\examples{
d_results <- cohens_d(drop_qcs(example_set))

}
