% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{combined_data}
\alias{combined_data}
\title{Retrieve both sample information and features}
\usage{
combined_data(object, assay.type = NULL)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{assay.type}{character, assay to be used in case of multiple assays}
}
\value{
A data frame with sample information plus all features as columns,
one row per sample.
}
\description{
Retrieve both sample information and features
}
\examples{
data(example_set)
combined_data(example_set)

}
