% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{correct_drift}
\alias{correct_drift}
\title{Correct drift using cubic spline}
\usage{
correct_drift(object, log_transform = TRUE, spar = NULL,
  spar_lower = 0.5, spar_upper = 1.5, check_quality = FALSE,
  condition = "RSD_r < 0 & D_ratio_r < 0", plotting = FALSE,
  file = NULL, width = 16, height = 8, color = "QC",
  shape = NULL, color_scale = NULL, shape_scale = NULL)
}
\arguments{
\item{object}{a MetaboSet object}

\item{spar}{smoothing parameter}

\item{spar_lower, spar_upper}{lower and upper limits for the smoothing parameter}

\item{condition}{a character specifying the condition used to decide whether drift correction
works adequately, see Details}

\item{plotting}{logical, whether plots should be drawn}

\item{file}{path to the PDF file where the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\value{
MetaboSet object as the one supplied, with drift corrected features
}
\description{
A wrapper function for applying cubic spline drift correction and saving
before and after plots
}
\details{
If \code{spar} is set to \code{NULL} (the default), the smoothing parameter will
be separately chosen for each feature from the range [\code{spar_lower, spar_upper}]
using cross validation. The \code{condition} parameter should be a character giving a condition compatible
with dplyr::filter. The condition is applied on the \strong{changes} in the quality metrics
RSD, RSD_r, D_ratio and D_ratio_r. For example, the default is "RSD_r < 0 and D_ratio_r < 0",
meaning that both RSD_r and D_ratio_r need to decrease in the drift correction, otherwise the
drift corrected feature is discarded and the original is retained. If \code{shape} is set to \code{NULL} (the default), the column used for color
is also used for shape
}
\seealso{
\code{\link{dc_cubic_spline}}, \code{\link[stats]{smooth.spline}} for details about the regression,
\code{\link{inspect_dc}} for analysing the drift correction results,
\code{\link{save_dc_plots}} for plotting the drift correction process for each feature
}
