% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{dc_cubic_spline}
\alias{dc_cubic_spline}
\title{Fit a cubic spline to correct drift}
\usage{
dc_cubic_spline(object, spar = NULL, spar_lower = 0.5,
  spar_upper = 1.5)
}
\arguments{
\item{object}{a MetaboSet object}

\item{spar}{smoothing parameter}

\item{spar_lower, spar_upper}{lower and upper limits for the smoothing parameter}
}
\value{
list with object = MetaboSet object as the one supplied, with drift corrected fetures
and predicted = matrix of the predicted values by the cubic spline (used in visualization)
}
\description{
Corrects the dirft in the features by applying smoothed cubic spline regression
to each feature separately.
}
\details{
If \code{spar} is set to \code{NULL} (the default), the smoothing parameter will
be separately chosen for each feature from the range [\code{spar_lower, spar_upper}]
using cross validation.
}
\seealso{
\code{\link[stats]{smooth.spline}} for details about the regression,
\code{\link{inspect_dc}} for analysing the drift correction results,
\code{\link{save_dc_plots}} for plotting the drift correction process for each feature
}
