% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{drop_qcs}
\alias{drop_qcs}
\title{Drop QC samples}
\usage{
drop_qcs(object)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}
}
\value{
A SummarizedExperiment object as the one supplied, 
without QC samples.
}
\description{
Drop QC samples
}
\examples{
data(toy_notame_set)
dim(toy_notame_set)
noqc <- drop_qcs(toy_notame_set)
dim(noqc)

}
