% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{drop_qcs}
\alias{drop_qcs}
\title{Drop QC samples}
\usage{
drop_qcs(object)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or \code{\link{MetaboSet}} object}
}
\value{
A SummarizedExperiment or MetaboSet object as the one supplied, 
without QC samples.
}
\description{
Drop QC samples
}
\examples{
data(example_set)
dim(example_set)
noqc <- drop_qcs(example_set)
dim(noqc)

}
