% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_set}
\alias{example_set}
\alias{hilic_neg_sample}
\alias{hilic_pos_sample}
\alias{rp_neg_sample}
\alias{rp_pos_sample}
\title{Toy data set}
\format{
An object of class \code{SummarizedExperiment} with 80 rows and 50 columns.

An object of class \code{SummarizedExperiment} with 20 rows and 50 columns.

An object of class \code{SummarizedExperiment} with 20 rows and 50 columns.

An object of class \code{SummarizedExperiment} with 20 rows and 50 columns.

An object of class \code{SummarizedExperiment} with 20 rows and 50 columns.
}
\usage{
example_set

hilic_neg_sample

hilic_pos_sample

rp_neg_sample

rp_pos_sample
}
\description{
Contains imaginary data used in testing the package functions. The dataset 
has 50 samples and 80 features. This dataset includes multiple observations 
from same subjects, sampled at two timepoints in separate batches and 
divided to two groups. The analytical modes are also available as separate 
SummarizedExperiment objects.
Note that across batches, the features don't have different feature ID's, 
m/z and retention time as would be the case with real-world data. In 
essence, the example data reflects that features were aligned perfectly 
between batches.
}
\keyword{datasets}
