% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{exponential}
\alias{exponential}
\alias{exponential,MetaboSet-method}
\alias{exponential,SummarizedExperiment-method}
\title{Exponential function}
\usage{
exponential(object, base = exp(1))

\S4method{exponential}{MetaboSet}(object, base = exp(1))

\S4method{exponential}{SummarizedExperiment}(object, base = exp(1))
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or \code{\link{MetaboSet}} object}

\item{base}{base of the exponential}
}
\value{
An object with altered feature abundances.
}
\description{
Apply the exponential function to peak table.
}
\examples{
data(example_set)
example_set <- mark_nas(example_set, value = 0)
log_data <- log(example_set)
orig_data <- exponential(log_data)

}
