% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{find_clusters}
\alias{find_clusters}
\title{Extract the densely connected clusters}
\usage{
find_clusters(connections, d_thresh = 0.8)
}
\arguments{
\item{connections}{data frame of pairs of signals that are linked together,
output of find_connections}

\item{d_thresh}{numeric, the minimum degree required for each signal in a cluster
expressed as a percentage of the maximum degree in the cluster}
}
\value{
a lis of clusters, each a list of:
\begin{itemize}
\item features: character vector of the names of the features included in the cluster
\item graph: an igraph object of the cluster
\end{itemize}
}
\description{
First forms clusters of compounds that are linked together. THen the clusters are pruned
so that in the final clusters, each feature is linked to at least a set percentage
of the other features in th e cluster.
}
