% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{find_connections}
\alias{find_connections}
\title{Find out which features are correlated within a specified retention time window}
\usage{
find_connections(
  data,
  features,
  corr_thresh = 0.9,
  rt_window = 1/60,
  name_col,
  mz_col,
  rt_col
)
}
\arguments{
\item{data}{data frame with the abundances of features, with features as columns}

\item{features}{data frame with feature information, fData(object)}

\item{corr_thresh}{numeric, the threshold of correlation to use in linking features}

\item{rt_window}{numeric, the retention time window to use in linking features. NOTE you
need to use the same unit as in the retention time column}

\item{name_col}{character, name of the column in features that contains feature names}

\item{mz_col}{character, name of the column in features that contains mass-to-charge ratios}

\item{rt_col}{character, name of the column in features that contains retention times}
}
\value{
a data frame of pairs of signals that are linked together
\itemize{
\item x & y: indexes and names of the signals
\item cor: correlation coefficient
\item mz_diff & rt_diff: mass and retention time difference
}
}
\description{
A part of the peak clustering algorithm. Iterates over all possible pairs of features
and records a connection between them if a) they have a Pearson correlation coefficient
higher than \code{corr_thresh} and b) their retention time difference is less than
\code{rt_window}
}
