% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{finish_log}
\alias{finish_log}
\title{Finish a log}
\usage{
finish_log()
}
\description{
Logs the current date and time and session info, and switches logging off.
}
\seealso{
\code{\link{init_log}}, \code{\link{log_text}}, \code{\link{log_state}}
}
