% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finite_helpers}
\alias{finite_helpers}
\alias{finite_sd}
\alias{finite_mean}
\alias{finite_median}
\alias{finite_mad}
\title{Summary statistics of finite elements}
\usage{
finite_sd(x)

finite_mean(x)

finite_median(x)

finite_mad(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
These functions first remove non-finite and missing values, then compute the summary statistic in question.
They are helper functions used for computing quality measurements.
}
