% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{flag}
\alias{flag}
\alias{flag<-}
\title{Get and set the values in the flag column}
\usage{
flag(object)

flag(object) <- value
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{value}{character vector, values for flag column}
}
\value{
Character vector of feature flags.

For the endomorphism, an object with a modified flag column.
}
\description{
Get and set the values in the flag column
}
\examples{
data(toy_notame_set)
# Get values in flag column of rowData
flag(toy_notame_set)

data(toy_notame_set)
# Flag a suspicious feature manually
flag(toy_notame_set)[1] <- "Contaminant, known from experience"
}
