% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{flag_contaminants}
\alias{flag_contaminants}
\title{Flag contaminants}
\usage{
flag_contaminants(
  object,
  blank_col,
  blank_label,
  flag_thresh = 0.05,
  flag_label = "Contaminant"
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{blank_col}{character, the column name in pData with blank labels}

\item{blank_label}{character, the label for blank samples in blank_col}

\item{flag_thresh}{numeric, the ratio threshold for flagging contaminants.
If the median of blanks > flag_thresh * median of biological samples, the feature gets flagged.}

\item{flag_label}{character, the label used when flagging contaminants. Can be changed if
sample processing contaminants and carryover contaminants are flagged separately.}
}
\description{
Flags contaminant features by comparing the median values of blanks and biological samples.
Biological sampels are defined as samples that are not marked as blanks and are not QCs.
If the median of blanks > the median of biological samples times a set ratio, the feature is
flagged as contaminant
}
