% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{flag_quality}
\alias{flag_quality}
\title{Flag low-quality features}
\usage{
flag_quality(object, assay.type = NULL, RSD = 0.2, D_ratio = 0.4)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{assay.type}{character, assay to be used in case of multiple assays}

\item{RSD}{character, minimum RSD to keep feature}

\item{D_ratio}{character, minimum D-ratio to keep feature}
}
\value{
a SummarizedExperiment object with the features flagged.
}
\description{
Flags low-quality features using the quality metrics defined in (Broadhurst 
2018). The metrics are described in more detain in Details. A dual RSD and 
D_ratio condition for keeping the features is given.
}
\details{
The quality metrics measure two things: internal spread of the QCs,
and spread of the QCs compared to the spread of the biological samples.
Internal spread is measured with relative standard deviation (RSD), also 
known as coefficient of variation (CV).
RSD used here is the non-parametric, robust version based on the 
median and median absolute deviation (MAD):
  \deqn{RSD_r = 1.4826 * MAD(QC) / median(QC)}
The spread of the QC samples compared to the biological samples is measured 
using a metric called D-ratio. The non-parametric, robust alternative is 
used here:
  \deqn{D_ratio_r = MAD(QC) / MAD(biological) }
The default condition keeps features that pass the following condition:
\deqn{RSD_r < 0.2 \& D_ratio_r < 0.4}
}
\examples{
data(toy_notame_set)
ex_set <- flag_quality(toy_notame_set)
rowData(ex_set)
# Custom condition
ex_set <- flag_quality(toy_notame_set, RSD = 0.1, D_ratio = 0.3)
rowData(ex_set)

}
\references{
Broadhurst, David et al. Guidelines and considerations for the 
use of system suitability and quality control samples in mass spectrometry 
assays applied in untargeted clinical metabolomic studies.
Metabolomics : Official journal of the Metabolomic Society vol. 14,6 (2018): 
72. doi:10.1007/s11306-018-1367-3
}
