% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{flag_report}
\alias{flag_report}
\title{A report of flagged features}
\usage{
flag_report(object)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}
}
\value{
A data frame with the number of features at each stage of flagging.
}
\description{
Computes the number of features at each stage of flagging for each mode.
}
\examples{
data(toy_notame_set)
flagged <- toy_notame_set |>
  mark_nas(0) |>
  flag_detection(group = "Group") |>
  flag_quality()
flag_report(flagged)

}
