% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fold_change}
\alias{fold_change}
\title{Fold change}
\usage{
fold_change(object, group = group_col(object))
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the group column}
}
\value{
data frame with fold changes for each feature
}
\description{
Computes fold change between each group for each feature.
}
\examples{
# Between groups
fc <- fold_change(example_set)
# Between time points
fc <- fold_change(example_set, group = "Time")

}
