% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{get_x}
\alias{get_x}
\title{A helper function for extracting predictor matrix with covariates}
\usage{
get_x(object, covariates)
}
\arguments{
\item{object}{a MetaboSet object}

\item{covariates}{character, column names of pData to use as covariates in the model, in addition to
molecular features}
}
\description{
A helper function for extracting predictor matrix with covariates
}
