% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{group_col}
\alias{group_col}
\alias{group_col<-}
\title{These generic functions access and set the special column for group labels}
\usage{
group_col(object)

group_col(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{string, name of column to be designated for holding group labels}
}
\description{
These generic functions access and set the special column for group labels
}
