% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{impute_rf}
\alias{impute_rf}
\title{Impute missing values using random forest}
\usage{
impute_rf(object, all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{...}{passed to MissForest function}
}
\value{
MetaboSet object as the one supplied, with missing values imputed.
}
\description{
Impute the missing values in the exprs part of the object using a
random forest. The estimated error in the imputation is logged.
It is recommended to ste the seed number for reproducibility
(it is called random forest for a reason).
This a wrapper around missForest::missForest.
Use parallelize = "variables" to run in parallel for faster testing.
NOTE: running in parallel prevents user from setting a seed number.
}
\seealso{
\code{\link[missForest]{missForest}} for detail about the algorithm and the parameters
}
