% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{init_log}
\alias{init_log}
\title{Initialize log to a file}
\usage{
init_log(log_file)
}
\arguments{
\item{log_file}{Path to the log file}
}
\description{
Initialize a log file with the current data and time.
All major operations run after this will be logged to the specified file.
}
\section{Warning}{

This overwrites the current contents of the file
}

\examples{
file_name <- "~/log.txt"
init_log(file_name)
# Print the contents of the file
scan(file_name, sep="\n", what = "chracter")

}
\seealso{
\code{\link{log_text}}, \code{\link{finish_log}}, \code{\link{log_state}}
}
