% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inverse_normalize}
\alias{inverse_normalize}
\title{Inverse-rank normalization}
\usage{
inverse_normalize(object, assay.type = NULL, name = NULL)
}
\arguments{
\item{object}{a SummarizedExperiment or MetaboSet object}

\item{assay.type}{character, assay to be used in case of multiple assays}

\item{name}{character, name of the resultant assay in case of multiple assays}
}
\value{
An object as the one supplied, with normalized features.
}
\description{
Applies inverse rank normalization to all features to approximate
a normal distribution.
}
\examples{
normalized <- inverse_normalize(example_set)
}
