% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{join_colData}
\alias{join_colData}
\alias{join_colData,SummarizedExperiment,data.frame-method}
\title{Join new columns to pheno data}
\usage{
join_colData(object, dframe)

\S4method{join_colData}{SummarizedExperiment,data.frame}(object, dframe)
}
\arguments{
\item{object}{a SummarizedExperiment object}

\item{dframe}{a data frame with the new information}
}
\value{
A SummarizedExperiment object with the new information added to 
colData(object).
}
\description{
Join a new data frame of information to pheno data of a SummarizedExperiment 
object.
}
\examples{
new_info <- data.frame(
  Sample_ID = colnames(example_set),
  BMI = stats::runif(ncol(example_set), 22, 26)
)
with_new_info <- join_colData(example_set, new_info)
colnames(colData(with_new_info))

}
