% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{join_results}
\alias{join_results}
\title{Join results to a MetaboSet object}
\usage{
join_results(object, dframe)
}
\arguments{
\item{object}{a MetaboSet object}

\item{dframe}{a dataframe with the results}
}
\value{
a MetaboSet object with the new information added to results(object)
}
\description{
Join a new data frame of results to a MetaboSet object. The data frame needs to have a column "Feature_ID".
This function is usually called on the results of one of the statistics functions.
}
\examples{
lm_results <- perform_lm(example_set, formula_char = "Feature ~ Group")
with_results <- join_results(example_set, lm_results)
colnames(results(with_results))

}
