% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log,MetaboSet-method}
\alias{log,MetaboSet-method}
\alias{log2,MetaboSet-method}
\alias{log10,MetaboSet-method}
\alias{log,SummarizedExperiment-method}
\alias{log2,SummarizedExperiment-method}
\alias{log10,SummarizedExperiment-method}
\title{Logarithm}
\usage{
\S4method{log}{MetaboSet}(x, base = exp(1))

\S4method{log2}{MetaboSet}(x)

\S4method{log10}{MetaboSet}(x)

\S4method{log}{SummarizedExperiment}(x, base = exp(1))

\S4method{log2}{SummarizedExperiment}(x)

\S4method{log10}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or \code{\link{MetaboSet}} object}

\item{base}{the base of the logarithm}
}
\value{
A SummarizedExperiment or MetaboSet object with peak table
transformed.
}
\description{
Log-transforms the peak table of SummarzedExperiment or MetaboSet object. 
Shortcuts for log2 and log10 also implemented.
For more information, see \code{\link{log}}.
}
