% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{mark_qcs}
\alias{mark_qcs}
\title{Replace NA values in pheno data columns with "QC"}
\usage{
mark_qcs(data, cols)
}
\arguments{
\item{data}{a data frame such as pheno_data returned by \code{\link{read_from_excel}}}

\item{cols}{the columns to fix}
}
\value{
a data frame with the column modified
}
\description{
Loops through specified columns in pheno data and replaces all NA values with "QC". Also transforms
the column to factor and sets "QC" as the last level.
}
\examples{
# Remove QC labels first
pheno_data <- pData(merged_sample)
pheno_data$Group[pheno_data$Group == "QC"] <- NA
head(pheno_data$Group, 20)
pheno_data <- mark_qcs(pheno_data, cols = "Group")
head(pheno_data$Group, 20)

}
