% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merge_batches}
\alias{merge_batches}
\title{Merge MetaboSet objects of batches together}
\usage{
merge_batches(...)
}
\arguments{
\item{...}{MetaboSet objects or a list of Metaboset objects}
}
\value{
A merged MetaboSet object
}
\description{
Merge MetaboSet objects of batches together
}
\examples{
batch1 <- merged_sample[, merged_sample$Batch == 1]
batch2 <- merged_sample[, merged_sample$Batch == 2]
merged <- merge_batches(batch1, batch2)

}
