% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merge_metabosets}
\alias{merge_metabosets}
\title{Merge MetaboSet objects together}
\usage{
merge_metabosets(...)
}
\arguments{
\item{...}{MetaboSet objects or a list of Metaboset objects}
}
\value{
A merged MetaboSet object
}
\description{
Merge MetaboSet objects together
}
