% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merge_metabosets}
\alias{merge_metabosets}
\title{Merge MetaboSet objects together}
\usage{
merge_metabosets(...)
}
\arguments{
\item{...}{MetaboSet objects or a list of Metaboset objects}
}
\value{
A merged MetaboSet object
}
\description{
Merge MetaboSet objects together
}
\examples{
merged <- merge_metabosets(hilic_neg_sample, hilic_pos_sample,
                           rp_neg_sample, rp_pos_sample)

}
