% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merge_metabosets}
\alias{merge_metabosets}
\title{Merge MetaboSet objects together}
\usage{
merge_metabosets(..., merge = c("features", "samples"))
}
\arguments{
\item{...}{\code{\link{MetaboSet}} objects or a list of MetaboSet objects}

\item{merge}{what to merge? features is used for combining analytical modes,
samples is used for batches}
}
\value{
A merged MetaboSet object.
}
\description{
Merges two or more MetaboSet objects together. Can be used to merge 
analytical modes or batches.
}
\details{
When merging samples, sample IDs that beging with "QC" or "Ref" are 
combined so that they have running numbers on them. This means that if both 
batches have samples called "QC_1", this will not result in an error,
but the sample IDs will be adjusted so that they are unique.
}
\examples{
# Merge analytical modes
data(hilic_neg_sample, hilic_pos_sample, rp_neg_sample, rp_pos_sample)
modes <- list(hilic_neg_sample, hilic_pos_sample, 
              rp_neg_sample, rp_pos_sample)
modes <- lapply(modes, function(mode) as(mode, "MetaboSet"))
merged <- merge_metabosets(modes)

# Merge batches
data(example_set)
batch1 <- example_set[, example_set$Batch == 1]
batch2 <- example_set[, example_set$Batch == 2]
merged <- merge_metabosets(batch1, batch2, merge = "samples")

}
