% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\docType{class}
\name{MetaboSet-class}
\alias{MetaboSet-class}
\alias{MetaboSet}
\alias{combined_data,MetaboSet-method}
\alias{group_col,MetaboSet-method}
\alias{time_col,MetaboSet-method}
\alias{subject_col,MetaboSet-method}
\alias{flag,MetaboSet-method}
\alias{join_fData,MetaboSet,data.frame-method}
\title{An S4 class used to represent LC-MS datasets}
\usage{
\S4method{combined_data}{MetaboSet}(object)

\S4method{group_col}{MetaboSet}(object)

\S4method{time_col}{MetaboSet}(object)

\S4method{subject_col}{MetaboSet}(object)

\S4method{flag}{MetaboSet}(object)

\S4method{join_fData}{MetaboSet,data.frame}(object, dframe)
}
\description{
MetaboSet is the main class used to represent data in the notame package.
It is built upon the \code{\link[Biobase]{ExpressionSet}} class from the Biobase
package. For more information, read the MetaboSet utility vignette.
In addition to the slots inherited from \code{\link[Biobase]{ExpressionSet}},
\code{MetaboSet} has three slots of its own. The extra three slots hold special
column names that are stored purely for convenience, as many functions use these as
defaults.
}
\section{Methods (by generic)}{
\itemize{
\item \code{combined_data}: sample information and features combined to a single data frame, one row per sample

\item \code{group_col}: access and set group_col

\item \code{time_col}: access and set time_col

\item \code{subject_col}: access and set subject_col

\item \code{flag}: access and set results

\item \code{join_fData}: join new information to fData
}}

\section{Slots}{

\describe{
\item{\code{group_col}}{character, name of the column holding group information}

\item{\code{time_col}}{character, name of the column holding time points}

\item{\code{subject_col}}{character, name of the column holding subject identifiers}
}}

