% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{muvr_analysis}
\alias{muvr_analysis}
\title{MUVR}
\usage{
muvr_analysis(
  object,
  y = NULL,
  id = NULL,
  multi_level = FALSE,
  multi_level_var = NULL,
  nRep = 5,
  nOuter = 6,
  nInner = nOuter - 1,
  varRatio = 0.75,
  method = c("PLS", "RF"),
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{y}{the target variable to predict}

\item{id}{subject ID variable in case of repeated measurements}

\item{multi_level}{whether multi-level modeling should be applied, see Details}

\item{multi_level_var}{the variable for splitting the data in multi-level modeling}

\item{nRep}{Number of repetitions of double CV, parameter of MUVR}

\item{nOuter}{Number of outer CV loop segments, parameter of MUVR}

\item{nInner}{Number of inner CV loop segments, parameter of MUVR}

\item{varRatio}{Ratio of variables to include in subsequent inner loop iteration,
parameter of MUVR}

\item{method}{Multivariate method. Supports 'PLS' and 'RF', parameter of MUVR}

\item{...}{other parameters to \code{MUVR::MUVR}}
}
\description{
A wrapper around the MUVR algorithm from the MUVR package. For more information
about the algorithm, visit https://gitlab.com/CarlBrunius/MUVR.
}
\examples{
\dontrun{
# Simple model, only 1 repetition for a quick example
rf_model <- muvr_analysis(drop_qcs(merged_sample), y = "Group", nRep = 1, method = "RF")

# PLS on multilevel variable
pls_model <- muvr_analysis(drop_qcs(example_set), multi_level = TRUE,
id = "Subject_ID", multi_level_var = "Time")
}

}
\seealso{
\code{\link[MUVR]{MUVR}}
}
