% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{mz_rt_plot}
\alias{mz_rt_plot}
\title{Plot m/z vs retention time plot}
\usage{
mz_rt_plot(
  object,
  p_col = NULL,
  p_limit = NULL,
  mz_col = NULL,
  rt_col = NULL,
  color = NULL,
  title = "m/z retention time",
  subtitle = NULL,
  color_scale = getOption("notame.color_scale_con"),
  ...
)
}
\arguments{
\item{object}{a MetaboSet object or a data frame. If x is a MetaboSet object, fData(x) is used.
If x is a data frame, it is used as is.}

\item{p_col}{the column name containing p-values. This is used to scale the size of the points.}

\item{p_limit}{numeric, limits plotted features by p-values. If NULL, plots all features.}

\item{mz_col, rt_col}{the column names for m/z and retention time. If NULL, automatic detection is attempted.}

\item{color}{the column name used to color the points}

\item{title}{The plot title}

\item{subtitle}{The plot subtitle}

\item{color_scale}{color scale as returned by a ggplot function. Defaults to current continuous color scale.}

\item{all_features}{logical, should all features be retained? Should be used only if x is a MetaboSet object.}
}
\value{
a ggplot object
}
\description{
Plots a scatter plot of results of statistical tests, where each point represents a feature.
The plot has retention time on x-axis, m/z on y-axis and the size of the points is scaled based on p-value
}
\examples{

# Compute results from a linear model
lm_results <- perform_lm(merged_sample, formula_char = "Feature ~ Group")
with_results <- join_fData(merged_sample, lm_results)

# Plot from the MetaboSet object
# automatically facet by analytical mode in variable Split
mz_rt_plot(with_results, p_col = "GroupB_P", color = "GroupB_Estimate")

# Plot the results from the results dataframe
mz_rt_plot(with_results, p_col = "GroupB_P", color = "GroupB_Estimate")

}
