% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{normalize_batches}
\alias{normalize_batches}
\title{Normalize batches}
\usage{
normalize_batches(object, batch, group = group_col(object), ref_label,
  ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch, group}{character, column names of pData with batch labels and group labels}

\item{ref_label}{the label of the reference group i.e. the group that is constant through batches}

\item{...}{additional parameters passed to batchCorr::normalizeBatches}
}
\value{
list, the object with normalized features and information on which features were corrected by ref samples in each batch.
}
\description{
Normalize bathces by either reference samples of population median.
Uses normalizeBatches function from the batchCorr package
}
