% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_paired_t_test}
\alias{perform_paired_t_test}
\alias{perform_pairwise_t_test}
\alias{perform_mann_whitney}
\alias{perform_wilcoxon_signed_rank}
\alias{perform_pairwise_non_parametric}
\title{Perform paired t-tests
DEPRECATED}
\usage{
perform_paired_t_test(object, group, id, all_features = FALSE, ...)

perform_pairwise_t_test(
  object,
  group = group_col(object),
  is_paired = FALSE,
  id = NULL,
  all_features = FALSE,
  ...
)

perform_mann_whitney(object, formula_char, all_features = FALSE, ...)

perform_wilcoxon_signed_rank(object, group, id, all_features = FALSE, ...)

perform_pairwise_non_parametric(
  object,
  group = group_col(object),
  is_paired = FALSE,
  id = NULL,
  all_features = FALSE,
  ...
)
}
\description{
Perform paired t-tests
DEPRECATED

Pairwise and paired t-test
DEPRECATED

Mann-Whitney u test
DEPRECATED

Wilcoxon signed rank test
DEPRECATED

Pairwise non-parametric tests
DEPRECATED
}
