% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{notame-package}
\alias{notame}
\alias{notame-package}
\title{\code{notame} package.}
\description{
Provides functionality for untargeted LC-MS metabolomics research as 
specified in the associated publication in the 'Metabolomics Data
Processing and Data Analysis—Current Best Practices' special issue of the 
Metabolites journal (2020). This includes data pretreatment, feature 
selection and supporting visualizations. Raw data preprocessing and 
functionality related to biological context, such as pathway analysis, is 
not included.
}
\details{
In roughly chronological order, the functionality of notame is as follows. 
Please see the vignettes and paper for more information.

Data pretreatment (for reducing unwanted variation and completing the 
dataset, returning modifed objects):
\itemize{
\item \code{\link{mark_nas}} mark specified values as missing
\item \code{\link{flag_detection}} flag features with low detection rate
\item \code{\link{flag_contaminants}} flag contaminants based on blanks
\item \code{\link{correct_drift}} correct drift using cubic spline
\item \code{\link{ruvs_qc}} Remove Unwanted Variation (RUV) between batches
\item \code{\link{pca_bhattacharyya_dist}} Bhattacharyya distance between 
batches in PCA space
\item \code{\link{perform_repeatability}} compute repeatability measures
\item \code{\link{assess_quality}} assess quality information of features
\item \code{\link{quality}} extract quality information of features
\item \code{\link{flag_quality}} flag low-quality features
\item \code{\link{flag_report}} report of flagged features
\item \code{\link{impute_rf}} impute missing values using random forest
\item \code{\link{impute_simple}} impute missing values using simple 
imputation
\item \code{\link{cluster_features}} cluster correlated features originating 
from the same metabolite
\item \code{\link{compress_clusters}} Compress clusters of features to a 
single feature
\item \code{\link{log}} logarithms
\item \code{\link{exponential}} exponential
\item \code{\link{pqn_normalization}} probabilistic quotient normalization
\item \code{\link{inverse_normalize}} inverse-rank normalization
\item \code{\link{scale}} scale
}

Data pretreatment visualizations (for visualization of data pretreatment and 
exploring the data at large, saved to file by default):
\itemize{
\item \code{\link{visualizations}} write all relevant data pretreatment 
visualizations to pdf
\item \code{\link{plot_injection_lm}} estimate the magnitude of drift
\item \code{\link{plot_sample_boxplots}} plot a boxplot for each sample
\item \code{\link{plot_dist_density}} plot distance density
\item \code{\link{plot_tsne}}, \code{\link{plot_pca}} t-SNE and PCA plot
\item \code{\link{plot_tsne_arrows}}, \code{\link{plot_pca_arrows}} t-SNE 
and PCA plots with arrows
\item \code{\link{plot_tsne_hexbin}}, \code{\link{plot_pca_hexbin}}
t-SNE and PCA hexbin plots
\item \code{\link{plot_dendrogram}} sample dendrogram
\item \code{\link{plot_sample_heatmap}} sample heatmap
\item \code{\link{plot_pca_loadings}} PCA loadings plot
\item \code{\link{plot_quality}} plot quality metrics
}

Feature selection – Univariate analysis (return data.frames):
\itemize{
\item \code{\link{perform_lm}} linear models 
\item \code{\link{perform_logistic}} logistic regression
\item \code{\link{perform_lmer}} linear mixed models 
\item \code{\link{perform_oneway_anova}} Welch’s ANOVA and classic ANOVA
\item \code{\link{perform_lm_anova}} linear models ANOVA table
\item \code{\link{perform_t_test}} pairwise and paired t-tests
\item \code{\link{perform_kruskal_wallis}} Kruskal-Wallis rank-sum test
\item \code{\link{perform_non_parametric}} pairwise and paired non-
parametric tests
\item \code{\link{perform_correlation_tests}} correlation test
\item \code{\link{perform_auc}} area under curve
\item \code{\link{perform_homoscedasticity_tests}} test homoscedasticity
\item \code{\link{cohens_d}} Cohen's D
\item \code{\link{fold_change}} fold change
\item \code{\link{summary_statistics}} summary statistics
\item \code{\link{summarize_results}} statistics cleaning
}

Feature selection – Supervised learning (return various objects):
\itemize{
\item \code{\link{muvr_analysis}} multivariate modelling with minimally 
biased variable selection (MUVR2) 
\item \code{\link{mixomics_pls}}, \code{\link{mixomics_plsda}} a simple 
PLS(-DA) model with set number of components and all features
\item \code{\link{mixomics_pls_optimize}},
\code{\link{mixomics_plsda_optimize}} test different numbers of components 
for PLS(-DA)
\item \code{\link{mixomics_spls_optimize}},
\code{\link{mixomics_splsda_optimize}} test different numbers of 
components and features for PLS(-DA)
\item \code{\link{fit_rf}}, \code{\link{importance_rf}} fit random forest
and feature importance
\item \code{\link{perform_permanova}} PERMANOVA
}

Feature-wise visualizations (these are often drawn for a subset of 
interesting features after analysis, saved by default):
\itemize{
\item \code{\link{save_beeswarm_plots}} save beeswarm plots of each feature 
by group
\item \code{\link{save_group_boxplots}} save box plots of each feature by 
group
\item \code{\link{save_scatter_plots}} save scatter plots of each feature 
against a set variable
\item \code{\link{save_subject_line_plots}} save line plots with mean
\item \code{\link{save_group_lineplots}} save line plots with errorbars by 
group
\item \code{\link{save_batch_plots}} save batch correction plots
}

Results visualizations (returned by default, save them using 
\code{\link{save_plot}}):
\itemize{
\item \code{\link{plot_p_histogram}} histogram of p-values
\item \code{\link{volcano_plot}} volcano plot
\item \code{\link{manhattan_plot}} manhattan plot
\item \code{\link{mz_rt_plot}} m/z vs retention time plot (cloud plot)
\item \code{\link{plot_effect_heatmap}} heatmap of effects between 
variables, such as correlations
}

Object utilities:
\itemize{
\item \code{\link{read_from_excel}} read formatted Excel files
\item \code{\link{construct_metabosets}} construct MetaboSet objects
\item \code{\link{write_to_excel}} write results to Excel file
\item \code{\link{group_col}} get and set name of the special column for 
group labels
\item \code{\link{time_col}} get and set the name of the special column for 
time points
\item \code{\link{subject_col}} get and set the name of the special column 
for subject identifiers
\item \code{\link{flag}} get and set values in the flag column
\item \code{\link{drop_flagged}} drop flagged features
\item \code{\link{drop_qcs}} drop QC samples
\item \code{\link{join_fData}} join new columns to feature data
\item \code{\link{join_pData}} join new columns to pheno data
\item \code{\link{combined_data}} retrieve both sample information and 
features
\item \code{\link{merge_objects}} merge SummarizedExperiment objects 
together
\item \code{\link{fix_object}} fix object for functioning of notame
}

Other utilities:
\itemize{
\item \code{\link{citations}} show citations
\item \code{\link{init_log}} initialize log to a file
\item \code{\link{log_text}} log text to the current log file
\item \code{\link{finish_log}} finish a log
\item \code{\link{save_plot}} save plot to chosen format
\item \code{\link{fix_MSMS}} transform the MS/MS output to publication ready
\item \code{\link{merge_metabosets}} merge MetaboSet objects together
}
}
\references{
Klåvus et al. (2020). "notame": Workflow for Non-Targeted LC-MS Metabolic 
Profiling. Metabolites, 10: 135.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanhineva-lab/notame}
  \item Report bugs at \url{https://github.com/hanhineva-lab/notame/issues}
}

}
\author{
\strong{Maintainer}: Vilhelm Suksi \email{vksuks@utu.fi} (\href{https://orcid.org/0009-0005-1108-518X}{ORCID})

Authors:
\itemize{
  \item Anton Klåvus (\href{https://orcid.org/0000-0003-2612-0230}{ORCID}) [copyright holder]
  \item Jussi Paananen (\href{https://orcid.org/0000-0001-5100-4907}{ORCID}) [copyright holder]
  \item Oskari Timonen (\href{https://orcid.org/0000-0002-6317-6260}{ORCID}) [copyright holder]
  \item Atte Lihtamo
  \item Retu Haikonen (\href{https://orcid.org/0000-0003-0830-3850}{ORCID})
  \item Leo Lahti (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
  \item Kati Hanhineva (\href{https://orcid.org/0000-0001-6834-7375}{ORCID})
}

Other contributors:
\itemize{
  \item Ville Koistinen (\href{https://orcid.org/0000-0003-1587-8361}{ORCID}) [contributor]
  \item Olli Kärkkäinen (\href{https://orcid.org/0000-0003-0825-4956}{ORCID}) [contributor]
  \item Artur Sannikov [contributor]
}

}
