% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{notame-package}
\alias{notame}
\alias{notame-package}
\title{\code{notame} package.}
\description{
Provides tabular data preprocessing and data wrangling functionality for 
untargeted LC-MS metabolomics research.
}
\references{
Kl책vus et al. (2020). "notame": Workflow for Non-Targeted LC-MS Metabolic 
Profiling. Metabolites, 10: 135.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanhineva-lab/notame}
  \item Report bugs at \url{https://github.com/hanhineva-lab/notame/issues}
}

}
\author{
\strong{Maintainer}: Vilhelm Suksi \email{vksuks@utu.fi} (\href{https://orcid.org/0009-0005-1108-518X}{ORCID})

Authors:
\itemize{
  \item Anton Kl책vus (\href{https://orcid.org/0000-0003-2612-0230}{ORCID}) [copyright holder]
  \item Jussi Paananen (\href{https://orcid.org/0000-0001-5100-4907}{ORCID}) [copyright holder]
  \item Oskari Timonen (\href{https://orcid.org/0000-0002-6317-6260}{ORCID}) [copyright holder]
  \item Atte Lihtamo
  \item Retu Haikonen (\href{https://orcid.org/0000-0003-0830-3850}{ORCID})
  \item Leo Lahti (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
  \item Kati Hanhineva (\href{https://orcid.org/0000-0001-6834-7375}{ORCID})
}

Other contributors:
\itemize{
  \item Ville Koistinen (\href{https://orcid.org/0000-0003-1587-8361}{ORCID}) [contributor]
  \item Olli K채rkk채inen (\href{https://orcid.org/0000-0003-0825-4956}{ORCID}) [contributor]
  \item Artur Sannikov [contributor]
}

}
