% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect.R
\name{pca_bhattacharyya_dist}
\alias{pca_bhattacharyya_dist}
\title{Bhattacharyya distance between bathces in PCA space}
\usage{
pca_bhattacharyya_dist(object, batch, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch}{column name of pData givinh the batch labels}
}
\value{
matrix of Bhattacharyya distances between batches
}
\description{
Computes Bhattacharyya distance between all pairs of batches after
projecting the samples into PCA space with pcaMethods::pca
}
