% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{pca_bhattacharyya_dist}
\alias{pca_bhattacharyya_dist}
\title{Bhattacharyya distance between bathces in PCA space}
\usage{
pca_bhattacharyya_dist(
  object,
  batch,
  all_features = FALSE,
  center = TRUE,
  scale = "uv",
  nPcs = 3,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch}{column name of pData givinh the batch labels}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before imputation.}

\item{center}{logical, should the data be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{nPcs}{the number of principal components to use}

\item{...}{other parameters to pcaMethods::pca}
}
\value{
matrix of Bhattacharyya distances between batches
}
\description{
Computes Bhattacharyya distance between all pairs of batches after
projecting the samples into PCA space with pcaMethods::pca
}
\examples{
# Batch correction
\dontrun{
batch_corrected <- normalize_batches(merged_sample, batch = "Batch", group = "QC", ref_label = "QC")
# Evaluate batch correction
pca_bhattacharyya_dist(merged_sample, batch = "Batch")
pca_bhattacharyya_dist(batch_corrected, batch = "Batch")
}
}
