% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_kruskal_wallis}
\alias{perform_kruskal_wallis}
\title{Perform Kruskal-Wallis Rank Sum Tests}
\usage{
perform_kruskal_wallis(
  object,
  formula_char,
  all_features = FALSE,
  separate_by = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{formula_char}{character, the formula to be used in the linear model (see Details)
Defaults to "Feature ~ group_col(object)}

\item{all_features}{should all features be included in FDR correction?}

\item{separate_by}{character, name of the column on which you want to separate the tests. Usually used when dealing
with multiple sample types in the same object.}
}
\value{
data frame with the results
}
\description{
Performs Kruskal-Wallis Rank Sum Test for equality
}
\details{
The model is fit on combined_data(object). Thus, column names
in pData(object) can be specified. To make the formulas flexible, the word "Feature"
must be used to signal the role of the features in the formula. "Feature" will be replaced
by the actual Feature IDs during model fitting. For example, if testing for equality of
means in study groups, use "Feature ~ Group".
}
\examples{
perform_kruskal_wallis(example_set, formula_char = "Feature ~ Group")
# Perform Kruskal-Wallis tests on Time points separately in each Group
perform_kruskal_wallis(drop_qcs(merged_sample), formula_char = "Feature ~ Time", separate_by = "Group")

}
\seealso{
\code{\link{kruskal.test}}
}
