% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_oneway_anova}
\alias{perform_oneway_anova}
\title{Perform ANOVA}
\usage{
perform_oneway_anova(
  object,
  formula_char,
  all_features = FALSE,
  separate_by = NULL,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{formula_char}{character, the formula to be used in the linear model (see Details)
Defaults to "Feature ~ group_col(object)}

\item{all_features}{should all features be included in FDR correction?}

\item{separate_by}{character, name of the column on which you want to separate the tests. Usually used when dealing
with multiple sample types in the same object.}

\item{...}{other parameters to \code{\link{oneway.test}}}
}
\value{
data frame with the results
}
\description{
Performs ANOVA with Welch's correction as default, to deal with heterogeneity of variances.
Can also perform classic ANOVA with assumption of equal variances.
Uses base R function \code{oneway.test}.
}
\details{
The model is fit on combined_data(object). Thus, column names
in pData(object) can be specified. To make the formulas flexible, the word "Feature"
must be used to signal the role of the features in the formula. "Feature" will be replaced
by the actual Feature IDs during model fitting. For example, if testing for equality of
means in study groups, use "Feature ~ Group".
}
\examples{
perform_oneway_anova(example_set, formula_char = "Feature ~ Group")

}
\seealso{
\code{\link{oneway.test}}
}
