% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_pairwise_t_test}
\alias{perform_pairwise_t_test}
\title{Perform pairwise t-tests}
\usage{
perform_pairwise_t_test(object, group = group_col(object),
  all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, column name of phenoData giving the groups}

\item{...}{other parameters passed to perform_t_test, and eventually to base R t.test}
}
\value{
data frame with the results
}
\description{
Performs pariowise t-tests between all study groups. NOTE! Does not use formula interface
}
\details{
P-values of each comparison are corrected separately from each other.
}
\seealso{
\code{\link{perform_t_test}}, \code{\link{t.test}}
}
