% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{perform_permanova}
\alias{perform_permanova}
\title{Perform PERMANOVA}
\usage{
perform_permanova(
  object,
  group,
  all_features = FALSE,
  transform = "Standardize columns",
  coef = "Pythagorean",
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{character, name of the column to compare}

\item{all_features}{should all features be included?}

\item{transform}{Transformation to use in \code{IniTransform}. By default uses "Standardize columns".}

\item{coef}{Coefficient to calculate continuous distances in \code{DistContinuous}.
By default uses Pythagorean distances.}

\item{...}{other parameters to \code{\link[PERMANOVA]{PERMANOVA}}}
}
\value{
PERMANOVA object
}
\description{
Performs permutational multivariate analysis of variance. Uses package called PERMANOVA.
}
\examples{
permanova_res <- perform_permanova(drop_qcs(example_set), group = "Group")

}
