% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{perform_repeatability}
\alias{perform_repeatability}
\title{Compute repeatability measures}
\usage{
perform_repeatability(object, group)
}
\arguments{
\item{object}{a MetaboSet object}

\item{group}{column name of pData givinh the group labels}
}
\value{
data frame with one row per feature with the repeatability measure
}
\description{
Computes repeatability for each feature with the following formula:
\deqn{\frac{\sigma^2_{between}}{\sigma^2_{between} + \sigma^2_{within}}}
The repeatability ranges from 0 to 1. Higher repeatability depicts less
variation between batches.
}
\examples{
# Batch correction
batch_corrected <- normalize_batches(merged_sample, batch = "Batch", group = "QC", ref_label = "QC")
# Evaluate batch correction
rep_orig <- perform_repeatability(merged_sample, group = "Group")
mean(rep_orig$Repeatability)
rep_corr <- perform_repeatability(batch_corrected, group = "Group")
mean(rep_corr$Repeatability)

}
