% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_wilcoxon_signed_rank}
\alias{perform_wilcoxon_signed_rank}
\title{Perform Wilcoxon signed rank test}
\usage{
perform_wilcoxon_signed_rank(object, group, id, all_features = FALSE, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{should all features be included in FDR correction?}

\item{...}{other parameters to \code{\link{wilcox.test}}}

\item{formula_char}{character, the formula to be used in the linear model (see Details)
Defaults to "Feature ~ group_col(object)}
}
\value{
data frame with the results
}
\description{
Performs Wilcoxon signed rank test
Uses base R function \code{wilcox.test} with \code{paired = TRUE}.
}
\details{
The model is fit on combined_data(object). Thus, column names
in pData(object) can be specified. To make the formulas flexible, the word "Feature"
must be used to signal the role of the features in the formula. "Feature" will be replaced
by the actual Feature IDs during model fitting. For example, if testing for equality of
medians in study groups, use "Feature ~ Group".
}
\examples{
perform_wilcoxon_signed_rank(drop_qcs(example_set), group = "Time", id = "Subject_ID")

}
\seealso{
\code{\link{wilcox.test}}
}
