% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Draw dendrograms}
\usage{
plot_dendrogram(object, all_features = FALSE,
  color = group_col(object), dist_method = "euclidean",
  clust_method = "ward.D2", center = TRUE, scale = "uv",
  title = "Dendrogram of hierarchical clustering", subtitle = NULL,
  color_scale = NULL)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{color}{character, name of the column used for coloring the sample labels}

\item{dist_method}{distance method used in clustering, see ?dist}

\item{clust_method}{method used in clustering, see ?hclust}

\item{center}{logical, should the data be centered?}

\item{scale}{scaling used, as in \code{pcaMethods::prep}. Default is "uv" for unit variance}

\item{title}{The plot title}

\item{subtitle}{The plot subtitle}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\value{
A ggplot object.
}
\description{
Draws a dendrogram of a hierarchical clustering applied to the samples of an experiment
}
\seealso{
\code{\link{dist}} \code{\link{hclust}}
}
