% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_dist_density}
\alias{plot_dist_density}
\title{Distance density plot}
\usage{
plot_dist_density(object, dist_method = "euclidean", fill_scale = NULL,
  title = NULL, subtitle = NULL)
}
\arguments{
\item{object}{a MetaboSet object}

\item{dist_method}{method for calculating the distances, passed to dist}

\item{fill_scale}{a scale for the fill of the density curves, as returned by a ggplot function}

\item{xlab}{the x}
}
\description{
Plot density of distances between samples in QC samples and actual samples
}
\seealso{
\code{\link[stats]{dist}}
}
