% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_dist_density}
\alias{plot_dist_density}
\title{Distance density plot}
\usage{
plot_dist_density(
  object,
  all_features = FALSE,
  dist_method = "euclidean",
  center = TRUE,
  scale = "uv",
  color_scale = getOption("notame.color_scale_dis"),
  fill_scale = getOption("notame.fill_scale_dis"),
  title = paste("Density plot of", dist_method, "distances between samples"),
  subtitle = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{dist_method}{method for calculating the distances, passed to dist}

\item{center}{logical, should the data be centered?}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{color_scale}{a scale for the color of the edge of density curves, as returned by a ggplot function}

\item{fill_scale}{a scale for the fill of the density curves, as returned by a ggplot function}

\item{title}{the plot title}

\item{subtitle}{the plot subtitle}
}
\value{
a ggplot object
}
\description{
Plot density of distances between samples in QC samples and actual samples
}
\examples{
plot_dist_density(merged_sample)
# Drift correction tightens QCs together
plot_dist_density(correct_drift(merged_sample))

}
\seealso{
\code{\link[stats]{dist}}
}
