% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_injection_lm}
\alias{plot_injection_lm}
\title{Estimate the magnitude of drift}
\usage{
plot_injection_lm(object, all_features = FALSE)
}
\arguments{
\item{object}{A MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}
}
\value{
A ggplot object
}
\description{
Plots histograms of p-values from linear regression model, where each feature is predicted
by injection order alone. The expected uniform distribution is represented by a dashed red line.
}
\seealso{
\code{\link{plot_p_histogram}}
}
