% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA scatter plot}
\usage{
plot_pca(
  object,
  pcs = c(1, 2),
  all_features = FALSE,
  center = TRUE,
  scale = "uv",
  color = group_col(object),
  shape = color,
  label = NULL,
  density = FALSE,
  title = "PCA",
  subtitle = NULL,
  color_scale = NA,
  shape_scale = getOption("notame.shape_scale"),
  fill_scale = getOption("notame.fill_scale_dis"),
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{pcs}{numeric vector of length 2, the principal components to plot}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{center}{logical, should the data be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{color}{character, name of the column used for coloring the points. Set to NULL for black color.}

\item{shape}{character, name of the column used for shape. Set to NULL for uniform round shapes.}

\item{label}{character, name of the column used for point labels}

\item{density}{logical, whether to include density plots to both axes. The density curves will be split and colored by the 'color' variable.}

\item{title, subtitle}{the titles of the plot}

\item{color_scale}{the color scale as returned by a ggplot function. Set to NA to choose the appropriate scale based on the class of the coloring variable.}

\item{shape_scale}{the shape scale as returned by a ggplot function}

\item{fill_scale}{the fill scale used for density curves. If a continuous variable is used as color, density curve will be colorless.}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
a ggplot object. If \code{density} is \code{TRUE}, the plot will consist of multiple
parts and is harder to modify.
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods and plots the two first principal components
\strong{CITATION:} When using this function, cite the \code{pcaMethods} package
}
\examples{
plot_pca(merged_sample, color = "Injection_order", shape = "Group")

}
\seealso{
\code{\link[pcaMethods]{pca}}
}
