% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA scatter plot}
\usage{
plot_pca(object, all_features = FALSE, center = TRUE, scale = "uv",
  color = group_col(object), shape = NULL, density = FALSE,
  title = "PCA", subtitle = NULL, color_scale = NULL,
  shape_scale = NULL, fill_scale = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{center}{logical, should the data be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{density}{logical, whether to include density plots to both axes}

\item{title, subtitle}{the titles of the plot}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{shape_scale}{the shape scale as returned by a ggplot function}

\item{fill_scale}{the fill scale used for density curves}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
a ggplot object. If \code{density} is \code{TRUE}, the plot will consist of multiple
parts and is harder to modify.
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods and plots the two first principal components
\strong{CITATION:} When using this function, cite the \code{pcaMethods} package
}
\seealso{
\code{\link[pcaMethods]{pca}}
}
