% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca_hexbin}
\alias{plot_pca_hexbin}
\title{PCA hexbin plot}
\usage{
plot_pca_hexbin(object, center = TRUE, scale = "uv", method = "ppca",
  fill = "Injection_order", summary_fun = "mean", bins = 10,
  title = "PCA", subtitle = NULL, fill_scale = NULL, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{center}{logical, should the data  be centered prior to PCA (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{method}{the method to use, see documentation in pcaMethods}

\item{fill}{character, name of the column used for coloring the hexagons}

\item{summary_fun}{the function used to compute the value for each hexagon}

\item{bins}{the number of bins in x and y axes}

\item{fill_scale}{the fill scale as returned by a ggplot function}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
A ggplot object.
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods and plots the two first principal components
}
\seealso{
\code{\link[pcaMethods]{pca}}
}
