% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot_pca_loadings}
\alias{plot_pca_loadings}
\title{PCA loadings plot}
\usage{
plot_pca_loadings(
  object,
  pcs = c(1, 2),
  all_features = FALSE,
  center = TRUE,
  scale = "uv",
  n_features = c(10, 10),
  title = "PCA loadings",
  subtitle = NULL,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{pcs}{numeric vector of length 2, the principal components to plot}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{center}{logical, should the data be centered prior to PCA? (usually yes)}

\item{scale}{scaling used, as in pcaMethods::prep. Default is "uv" for unit variance}

\item{n_features}{numeric vector of length two, number of top feature to plot
for each principal component}

\item{title, subtitle}{the titles of the plot}

\item{...}{additional arguments passed to pcaMethods::pca}
}
\value{
a ggplot object.
}
\description{
Computes PCA using one of the methods provided in the Bioconductor package
pcaMethods and plots the loadings of first principal components
\strong{CITATION:} When using this function, cite the \code{pcaMethods} package
}
\examples{
plot_pca_loadings(merged_sample, n_features = c(2, 4))

}
\seealso{
\code{\link[pcaMethods]{pca}}
}
