% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualizations.R
\name{plot_quality}
\alias{plot_quality}
\title{Plot quality metrics}
\usage{
plot_quality(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\value{
a ggplot object
}
\description{
Plots distribution of each quality metric, and a distribution of the flags.
}
\examples{
plot_quality(example_set)

}
