% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{quality}
\alias{quality}
\title{Extract quality information of features}
\usage{
quality(object)
}
\arguments{
\item{object}{a MetaboSet object}
}
\description{
Extract quality information of features
}
