% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{quality}
\alias{quality}
\title{Extract quality information of features}
\usage{
quality(object)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}
}
\value{
A data frame with quality metrics for each feature.
}
\description{
Extract quality information of features
}
\examples{
data(toy_notame_set)
ex_set <- assess_quality(toy_notame_set)
quality(ex_set)

}
