% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{ruvs_qc}
\alias{ruvs_qc}
\title{Remove Unwanted Variation}
\usage{
ruvs_qc(object, batch, replicates, k = 3, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{batch}{the column name for batch labels}

\item{replicates}{list of numeric vectors, indexes of replicates}

\item{k}{The number of factors of unwanted variation to be estimated from the data.}

\item{...}{other parameters passed to RUVSeq::RUVs}
}
\value{
a MetaboSet object with the normalized data
}
\description{
An interface for the RUVs method in RUVSeq package.
}
