% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_beeswarm_plots}
\alias{save_beeswarm_plots}
\title{Save beeswarm plots of each feature by group}
\usage{
save_beeswarm_plots(
  object,
  all_features = FALSE,
  file,
  width = 8,
  height = 6,
  x = group_col(object),
  add_boxplots = FALSE,
  color = group_col(object),
  color_scale = NULL
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{file}{path to the PDF file where the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{x}{character, name of the column to be used as x-axis}

\item{add_boxplots}{logical, should boxplots be added to the figure?}

\item{color}{character, name of the column to be used for coloring}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\description{
Draws a beeswarm plot of feature abundances in each group.
A separate plot is drawn for each feature.
}
\examples{
\dontrun{
# Default beeswarms by group
save_beeswarm_plots(drop_qcs(merged_sample), file = "beeswarms.pdf")
# x and color can be a different variable
save_beeswarm_plots(drop_qcs(merged_sample), file = "beeswarms_time.pdf",
                    x = "Time", color = "Group")
}

}
