% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_beeswarm_plots}
\alias{save_beeswarm_plots}
\title{Save beeswarm plots of each feature by group}
\usage{
save_beeswarm_plots(
  object,
  all_features = FALSE,
  prefix,
  format = "emf",
  x = group_col(object),
  add_boxplots = FALSE,
  title = "Feature_ID",
  subtitle = NULL,
  color = group_col(object),
  color_scale = NULL,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{prefix}{character, a file path prefix added to the file paths}

\item{format}{character, format in which the plots should be saved}

\item{x}{character, name of the column to be used as x-axis}

\item{add_boxplots}{logical, should boxplots be added to the figure?}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{color}{character, name of the column to be used for coloring}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{...}{other arguments to save_plot function like width and height}
}
\description{
Draws a beeswarm plot of feature abundances in each group.
A separate plot is drawn and saved for each feature.
}
\examples{
\dontrun{
# Default beeswarms by group
save_beeswarm_plots(drop_qcs(merged_sample),
                    prefix = "./beeswarm_plots/",
                    format = "pdf"
                    )
# x and color can be a different variable
save_beeswarm_plots(drop_qcs(merged_sample),
                    prefix = "./beeswarm_plots/",
                    format = "pdf",
                    x = "Time",
                    color = "Group")
}

}
\seealso{
\code{\link[notame]{save_plot}}
}
