% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_beeswarm_plots}
\alias{save_beeswarm_plots}
\title{Save beeswarm plots of each feature by group}
\usage{
save_beeswarm_plots(
  object,
  all_features = FALSE,
  save = TRUE,
  file_path = NULL,
  format = "emf",
  x = group_col(object),
  add_boxplots = FALSE,
  title = "Feature_ID",
  subtitle = NULL,
  color = group_col(object),
  color_scale = getOption("notame.color_scale_dis"),
  text_base_size = 14,
  cex = 2,
  size = 2,
  title_line_length = 40,
  theme = theme_bw(base_size = text_base_size),
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{file_path}{character, a file path for PDF or prefix added to the file paths for other formats}

\item{format}{character, format in which the plots should be saved}

\item{x}{character, name of the column to be used as x-axis}

\item{add_boxplots}{logical, should boxplots be added to the figure?}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{color}{character, name of the column to be used for coloring}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{text_base_size}{integer, base size for text in figures}

\item{cex}{numeric, scaling for adjusting point spacing}

\item{size}{numeric, size of points}

\item{title_line_length}{integer, maximum length of the title line in characters, passed to stringr::str_wrap}

\item{theme}{a ggplot theme to be added to the plot}

\item{...}{other arguments to graphic device functions, like width and height}
}
\description{
Draws a beeswarm plot of feature abundances in each group.
A separate plot is drawn and saved for each feature.
}
\examples{
\dontrun{
# Default beeswarms by group
save_beeswarm_plots(drop_qcs(merged_sample),
  file_path = "./beeswarm_plots.pdf",
  format = "pdf"
)
# x and color can be a different variable
save_beeswarm_plots(drop_qcs(merged_sample)[1:10],
  file_path = "./beeswarm_plots/",
  format = "png",
  x = "Time",
  color = "Group"
)
}
# Plot one feature
save_beeswarm_plots(drop_qcs(merged_sample)[5, ], save = FALSE)
}
\seealso{
\code{\link[notame]{save_plot}}
}
