% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_correction.R
\name{save_dc_plots}
\alias{save_dc_plots}
\title{Drift correction plots}
\usage{
save_dc_plots(
  orig,
  dc,
  predicted,
  file,
  log_transform = TRUE,
  width = 16,
  height = 8,
  color = "QC",
  shape = color,
  color_scale = getOption("notame.color_scale_dis"),
  shape_scale = scale_shape_manual(values = c(15, 16))
)
}
\arguments{
\item{orig}{a MetaboSet object, before drift correction}

\item{dc}{a MetaboSet object, after drift correction as returned by correct_drift}

\item{predicted}{a matrix of predicted values, as returned by dc_cubic_spline}

\item{file}{path to the PDF file where the plots should be saved}

\item{log_transform}{logical, was the drift correction done on log-transformed data?}

\item{width, height}{width and height of the plots in inches}

\item{color}{character, name of the column used for coloring the points}

\item{shape}{character, name of the column used for shape}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{shape_scale}{the shape scale as returned by a ggplot function}
}
\description{
Plots the data before and after drift correction, with the regression line drawn with
the original data. If the drift correction was done on log-transformed data, then
plots of both the original and log-transformed data before and after correction are drawn.
The plot shows 2 standard deviation spread for both QC samples and regular samples.
}
\details{
If \code{shape} is set to \code{NULL} (the default), the column used for color
is also used for shape
}
\examples{
\dontrun{
dc <- dc_cubic_spline(merged_sample)
corrected <- dc$object
inspected <- inspect_dc(orig = merged_sample, dc = corrected,
                        check_quality = TRUE)
save_dc_plots(orig = merged_sample, dc = corrected, predicted = dc$predicted,
              file = "drift_plots.pdf")
}
}
\seealso{
\code{\link{correct_drift}}, \code{\link{inspect_dc}}
}
