% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_feature_plots}
\alias{save_feature_plots}
\title{Save plots of individual features}
\usage{
save_feature_plots(
  object,
  file_path,
  format,
  title,
  subtitle,
  text_base_size,
  plot_fun,
  ...
)
}
\arguments{
\item{plot_fun}{a function with arguments:
data frame from combined_data(object)
feature id
Should return a ggplot object for plotting}

\item{...}{arguments to code{\link[ggplot2]{ggsave}}}
}
\description{
Helper function for saving plots of individual features
to either one multi-page PDF or separate EMF figures
}
