% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_group_boxplots}
\alias{save_group_boxplots}
\title{Save box plots of each feature by group}
\usage{
save_group_boxplots(
  object,
  all_features = FALSE,
  prefix,
  format = "emf",
  width = 8,
  height = 6,
  res = 300,
  x = group_col(object),
  color = group_col(object),
  title = "Feature_ID",
  subtitle = NULL,
  color_scale = getOption("notame.color_scale_dis")
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{prefix}{character, a file path prefix added to the file paths}

\item{format}{character, format in which the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{res}{integer, resolution in ppi for non-vector formats}

\item{x}{character, name of the column to be used as x-axis}

\item{color}{character, name of the column to be used for coloring}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{color_scale}{the color scale as returned by a ggplot function}
}
\description{
Draws a boxplot of feature abundances in each group.
A separate plot is drawn and saved for each feature.
}
\examples{
\dontrun{
# Default boxplots by group
save_group_boxplots(drop_qcs(merged_sample),
                    prefix = "./group_boxplots/",
                    format = "pdf"
                    )
# x and color can be a different variable
save_group_boxplots(drop_qcs(merged_sample),
                    prefix = "./time_boxplots/",
                    format = "pdf",
                    x = "Time",
                    color = "Group"
                    )
}

}
\seealso{
\code{\link[notame]{save_plot}}
}
