% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_group_boxplots}
\alias{save_group_boxplots}
\title{Save box plots of each feature by group}
\usage{
save_group_boxplots(
  object,
  all_features = FALSE,
  save = TRUE,
  file_path = NULL,
  format = "emf",
  x = group_col(object),
  color = group_col(object),
  title = "Feature_ID",
  subtitle = NULL,
  color_scale = getOption("notame.color_scale_dis"),
  text_base_size = 14,
  box_width = 0.8,
  line_width = 0.5,
  point_size = 3,
  theme = theme_bw(base_size = text_base_size),
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default),
flagged features are removed before visualization.}

\item{save}{logical, if false, the plots are not saved but returned as a list}

\item{file_path}{character, a file path for PDF or prefix added to the file paths for other formats}

\item{format}{character, format in which the plots should be saved}

\item{x}{character, name of the column to be used as x-axis}

\item{color}{character, name of the column to be used for coloring}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{text_base_size}{integer, base size for text in figures}

\item{box_width}{numeric, width of the boxes}

\item{line_width}{numeric, width of the lines}

\item{point_size}{numeric, size of the mean points}

\item{theme}{a ggplot theme to be added to the plot}

\item{...}{other arguments to graphic device functions, like width and height}
}
\description{
Draws a boxplot of feature abundances in each group.
A separate plot is drawn and saved for each feature.
}
\examples{
\dontrun{
# Default boxplots by group
save_group_boxplots(drop_qcs(merged_sample),
  file_path = "./group_boxplots.pdf",
  format = "pdf", title = NULL
)
# x and color can be a different variable
save_group_boxplots(drop_qcs(merged_sample)[1:10],
  file_path = "./time_boxplots/",
  format = "emf",
  x = "Time",
  color = "Group", title = NULL
)
}
# Plot one feature
save_group_boxplots(drop_qcs(merged_sample)[5, ], save = FALSE)
}
\seealso{
\code{\link[notame]{save_plot}}
}
