% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_group_lineplots}
\alias{save_group_lineplots}
\title{Save line plots with errorbars by group}
\usage{
save_group_lineplots(
  object,
  all_features = FALSE,
  file_path,
  format = "emf",
  x = time_col(object),
  group = group_col(object),
  title = "Feature_ID",
  subtitle = NULL,
  fun.data = "mean_cl_boot",
  fun = NULL,
  fun.min = NULL,
  fun.max = NULL,
  position_dodge_amount = 0.2,
  color_scale = getOption("notame.color_scale_dis"),
  text_base_size = 14,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{file_path}{character, a file path for PDF or prefix added to the file paths for other formats}

\item{format}{character, format in which the plots should be saved}

\item{x}{character, name of the column to be used as x-axis}

\item{group}{character, name of the column containing group information, used for coloring}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{fun.data}{passed to ggplot2::stat_summary and used for errorbars,
"A function that is given the complete data and should return a data frame with variables ymin, y, and ymax."}

\item{fun.min, fun, fun.max}{Alternative to fun.data, passed to ggplot2::stat_summary,
"supply three individual functions that are each passed a vector of x's and should return a single number"}

\item{position_dodge_amount}{numeric: how much the group mean points should dodge away from each other}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{text_base_size}{integer, base size for text in figures}

\item{...}{other arguments to graphic device functions, like width and height}
}
\description{
Plots the change in the feature abundances as a function of e.g. time.
A line is drawn for each group and error bars are added.
A separate plot is drawn for each feature.
}
\examples{
\dontrun{
save_group_lineplots(drop_qcs(merged_sample),
                    file_path = "./group_line_plots.pdf",
                    format = "pdf"
)
save_group_lineplots(drop_qcs(merged_sample)[1:10],
                    file_path = "./group_line_plots/",
                    format = "png"
)
}

}
\seealso{
\code{\link[notame]{save_plot}},
\code{\link[ggplot2]{stat_summary}}
}
