% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utils.R
\name{save_plot}
\alias{save_plot}
\title{Save plot to PDF}
\usage{
save_plot(p, file, ...)
}
\arguments{
\item{p}{a ggplot object}

\item{file}{the file path}

\item{...}{other arguments to pdf, like width and height}
}
\description{
Saves the given plot to a PDF file. If an error occurs with the plot, an empty file is created.
}
\seealso{
\code{\link[grDevices]{pdf}}
}
