% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_scatter_plots}
\alias{save_scatter_plots}
\title{Save scatter plots of each feature against a set variable}
\usage{
save_scatter_plots(
  object,
  x = group_col(object),
  file,
  width = 8,
  height = 6,
  all_features = FALSE,
  color = NULL,
  color_scale = NA,
  shape = NULL,
  title = "Feature_ID",
  subtitle = NULL,
  shape_scale = getOption("notame.shape_scale")
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{x}{character, name of the column to be used as x-axis}

\item{file}{path to the PDF file where the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{all_features}{logical, should all features be used? If FALSE
(the default), flagged features are removed before visualization.}

\item{color}{character, name of the column to be used for coloring}

\item{color_scale}{the color scale as returned by a ggplot function.
Set to NA to choose the appropriate scale based on the class of the coloring variable.}

\item{shape}{character, name of the column used for shape}

\item{title, subtitle}{column names from fData to use as plot title and subtitle. Set to NULL for no title/subtitle}

\item{shape_scale}{the shape scale as returned by a ggplot function}
}
\description{
Draws a scatterplots with a feature on y-axis and another variable on x-axis.
A separate plot is drawn for each feature.
}
\examples{
\dontrun{
# Against injection order, colored by group
save_scatter_plots(object = merged_sample,
                           x = "Injection_order",
                           color = "Group",
                           file = "injection_scatter.pdf")
}

}
