% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_subject_line_plots}
\alias{save_subject_line_plots}
\title{Save line plots with mean}
\usage{
save_subject_line_plots(object, all_features = FALSE, file, width = 8,
  height = 6, x = time_col(object), id = subject_col(object),
  color = NA, color_scale = NULL, facet = NULL)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{file}{path to the PDF file where the plots should be saved}

\item{width, height}{width and height of the plots in inches}

\item{x}{character, name of the column to be used as x-axis}

\item{id}{character, name of the column containing subject IDs}

\item{color}{character, the column name to color the lines by (optional)}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{facet}{character, the column name to facet by (optional, usually same as color)}
}
\description{
Plots the change in the feature abundances as a function of e.g. time.
A line is drawn for each subject and a mean line is added.
A separate plot is drawn for each feature.
}
\examples{
\dontrun{save_subject_line_plots(drop_qcs(example_set), file = "subject_lines.pdf")}

}
