% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{save_subject_line_plots}
\alias{save_subject_line_plots}
\title{Save line plots with mean}
\usage{
save_subject_line_plots(
  object,
  all_features = FALSE,
  file_path,
  format = "emf",
  x = time_col(object),
  id = subject_col(object),
  title = "Feature_ID",
  subtitle = NULL,
  color = NA,
  color_scale = getOption("notame.color_scale_dis"),
  facet = NULL,
  ...
)
}
\arguments{
\item{object}{a MetaboSet object}

\item{all_features}{logical, should all features be used? If FALSE (the default), flagged features are removed before visualization.}

\item{file_path}{character, a file path for PDF or prefix added to the file paths for other formats}

\item{format}{character, format in which the plots should be saved}

\item{x}{character, name of the column to be used as x-axis}

\item{id}{character, name of the column containing subject IDs}

\item{title, subtitle}{column names from fData to use as plot title/filename and subtitle.
Set to NULL for no title/subtitle, this creates running numbered filenames}

\item{color}{character, the column name to color the lines by (optional)}

\item{color_scale}{the color scale as returned by a ggplot function}

\item{facet}{character, the column name to facet by (optional, usually same as color)}

\item{...}{other arguments to graphic device functions, like width and height}
}
\description{
Plots the change in the feature abundances as a function of e.g. time.
A line is drawn for each subject and a mean line is added.
A separate plot is drawn and saved for each feature.
}
\examples{
\dontrun{save_subject_line_plots(drop_qcs(example_set),
                    file_path = "./subject_line_plots.pdf",
                    format = "pdf"
                    )}

}
\seealso{
\code{\link[notame]{save_plot}}
}
