% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{scale}
\alias{scale}
\alias{scale,MetaboSet-method}
\alias{scale,SummarizedExperiment-method}
\title{Scale peak table}
\usage{
\S4method{scale}{MetaboSet}(x, center = TRUE, scale = TRUE)

\S4method{scale}{SummarizedExperiment}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
or \code{\link{MetaboSet}} object}

\item{center, scale}{as in base scale function}
}
\value{
A SummarizedExperiment or MetaboSet object with modified peak table.
}
\description{
Applies the base R function scale to transposed peak table. 
See \code{\link[base]{scale}} for details.
}
\examples{
scaled_set <- scale(example_set)
}
