% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{subject_col}
\alias{subject_col}
\alias{subject_col<-}
\title{These generic functions access and set the special column for subject identifiers}
\usage{
subject_col(object)

subject_col(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{string, name of column to be designated for holding subject identifiers}
}
\description{
These generic functions access and set the special column for subject identifiers
}
\examples{
# Get name of subject variable
subject_col(example_set)
# Set subject variable
subject_col(example_set) <- "Subject_ID"
}
