% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{time_col}
\alias{time_col}
\alias{time_col<-}
\title{Get and set the name of the special column for time points}
\usage{
time_col(object)

time_col(object) <- value
}
\arguments{
\item{object}{a MetaboSet object}

\item{value}{string, name of column to be designated for holding time points}
}
\value{
Character, name of time variable.

For the endomorphism, an object with the time variable set to the 
specified variable.
}
\description{
Get and set the name of the special column for time points
}
\examples{
# Get name of time variable
time_col(example_set)

# Set time variable
time_col(example_set) <- "Time"

}
