% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_clustering.R
\name{visualize_clusters}
\alias{visualize_clusters}
\title{Visualize clusters of fetaures}
\usage{
visualize_clusters(data, features, clusters, min_size, rt_window, name_col,
  mz_col, rt_col, file_path)
}
\description{
Draws multiple visualizations of each cluster, creating a separate file for each cluster
}
