% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(data, effect, p, log2_effect = FALSE,
  center_x_axis = TRUE, title = "Volcano plot", subtitle = NA, ...)
}
\arguments{
\item{data}{a data frame with the effect size and p-values}

\item{effect, p}{the column names of effect size and p-values}

\item{log2_effect}{logical, whether effect size should be plotted on a log2 axis}

\item{center_x_axis}{logical, whether x-axis should be centered. If \code{TRUE}, the "zero-effect" will
be on the middle of the plot. The "zero effect" is 0 if \code{log2_effect = FALSE} and 1 if  \code{log2_effect = TRUE}}

\item{title, subtitle}{the title and subtitle of the plot}

\item{...}{parameters passed to \code{\link[ggplot2]{geom_point}}, such as shape and alpha values.}
}
\value{
a ggplot object
}
\description{
Draws a volcano plot of effect size and p-values.
}
