% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{write_to_excel}
\alias{write_to_excel}
\title{Write results to Excel file}
\usage{
write_to_excel(object, file, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{file}{path to the file to write}

\item{...}{Additional parameters passed to \code{openxlsx::write.xlsx}}
}
\description{
Writes all the data in a MetaboSet object to an Excel spreadsheet.
The format is similar to the one used to read data in, except for the fact that
EVERYTHING NEEDS TO BE WRITTEN AS TEXT. To fix numeric values in Excel,
choose any cell with a number, press Ctrl + A, then go to the dropdown menu
in upper left corner and choose "Convert to Number". This will fix the file,
but can take quite a while.
}
