% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{write_to_excel}
\alias{write_to_excel}
\title{Write results to Excel file}
\usage{
write_to_excel(object, file, ...)
}
\arguments{
\item{object}{a MetaboSet object}

\item{file}{path to the file to write}

\item{...}{Additional parameters passed to \code{openxlsx::write.xlsx}}
}
\description{
Writes all the data in a MetaboSet object to an Excel spreadsheet.
The format is similar to the one used to read data in, except that
the results from statistics are added to the right.
}
